### 2016-08-04 #47 @whedon
[REVIEW]: A Structured-Light Scanning Software for Rapid Geometry Acquisition
**Submitting author:** @charalambos (<a href="http://orcid.org/0000-0001-5666-5026">Charalambos Poullis</a>)
**Repository:** <a href="https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU" target ="_blank">https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU</a>
**Version:** v4.0.0
