### 2016-08-05 7ca3f07
Add Open Journal Badge
