### 2016-08-05 36a1a53
Raw paper sources added
