### 2016-08-05 7e5f98f
Moved images out of image folder
