+++ doc/paper/elsarticle-template.tex
@@ -149,7 +149,7 @@ The paper is organized as follows: Section \ref{sec:related_work} provides a bri
 
 \begin{figure*}[!ht]
  \centering
- \includegraphics[scale=0.5]{./images/systemOverview.png}
+ \includegraphics[scale=0.5]{./systemOverview.png}
  \caption{\label{fig:system_overview} System Overview}
 \end{figure*}
 
@@ -186,7 +186,7 @@ An example of the process is shown in Figure \ref{fig:pixel_encoding} where the 
 
 \begin{figure}[!ht]
  \centering
- \includegraphics[scale=0.8]{./images/pixelSeq.png}
+ \includegraphics[scale=0.8]{./pixelSeq.png}
  \caption{\label{fig:pixel_encoding} Encoding of a single pixel $p$ at location $(209,660)$ is encoded into two sequences of $1x1$ patterns in temporal domain; one sequence representing the column and the other the row information.}
 \end{figure}
 
@@ -196,7 +196,7 @@ The result is a sequence of binary encoded patterns where each pattern represent
 
 \begin{figure}[!ht]
  \centering
- \includegraphics[scale=0.5]{./images/32x32binSeq.png}
+ \includegraphics[scale=0.5]{./32x32binSeq.png}
  \caption{\label{fig:pattern_encoding} Encoding the information into a sequence of patterns in temporal domain.}
 \end{figure}
 
