### 2016-08-05 e360e0a
edited tex file to remove images folder
