+++ doc/paper/elsarticle-template.tex
@@ -233,8 +233,8 @@ Gray codes can be calculated by first computing the Binary representation of a n
 
 \begin{figure*}[!ht]
 	\centering
-	\subfloat[\label{fig:bin2gray}]{\includegraphics[scale=0.75]{./images/binary_to_Gray.png}}
-	\subfloat[\label{fig:gray2bin}]{	\includegraphics[scale=0.75]{./images/gray_to_bin.png}}
+	\subfloat[\label{fig:bin2gray}]{\includegraphics[scale=0.75]{./binary_to_Gray.png}}
+	\subfloat[\label{fig:gray2bin}]{	\includegraphics[scale=0.75]{./gray_to_bin.png}}
 	\caption{(a) Encoding of decimal number 93. The Binary Code representing the decimal number 93 is converted into Gray Code. (b) Conversion of Gray Code to Binary Code.}
 \end{figure*}
 
@@ -243,9 +243,9 @@ copy the most significant bit as is and for the remaining bits (taking one bit a
 
 \begin{figure}[!ht]
 	\centering
-	\subfloat[\label{fig:bin}]{\includegraphics[scale=0.4]{./images/binary.png}}
+	\subfloat[\label{fig:bin}]{\includegraphics[scale=0.4]{./binary.png}}
 	\\
-	\subfloat[\label{fig:gray}]{\includegraphics[scale=0.4]{./images/gray.png}}
+	\subfloat[\label{fig:gray}]{\includegraphics[scale=0.4]{./gray.png}}
 	\caption{(a) Binary encoding of the columns for an area of 8x8 pixels. (b) Gray encoding corresponding to the same columns of the area in (a).}
 \end{figure}
 
@@ -269,7 +269,7 @@ When considering a camera, the image is formed when light rays being reflected i
 
 \begin{figure}[!ht]
 	\centering
-	\includegraphics[scale=0.8]{./images/rays_and_camera.png}
+	\includegraphics[scale=0.8]{./rays_and_camera.png}
 	\caption{\label{fig:camRays} Image formation. Light rays being reflected in the scene, pass through the camera lens, then through the aperture and interact with the photo-sensitive light sensor.}
 \end{figure}
 
@@ -317,7 +317,7 @@ Although, theoretically the camera matrix $C$ can be computed from a single imag
 
 \begin{figure}[!ht]
 	\centering
-	\includegraphics[scale=0.6]{./images/extrCalib.png}
+	\includegraphics[scale=0.6]{./extrCalib.png}
 	\caption{\label{fig:calibration} Multiple Camera Calibration. For each orientation and position of the calibration board, images are captured from all cameras. This allows the use of a common reference point i.e. world origin.}
 \end{figure}
 
@@ -340,11 +340,11 @@ An acquisition set consists of the images captured by each camera for each patte
 
 \begin{figure}[!ht]
 	\centering
-	\subfloat[\label{fig:view1}]{\includegraphics[scale=0.10]{./images/half_artemis01.jpg}}
+	\subfloat[\label{fig:view1}]{\includegraphics[scale=0.10]{./half_artemis01.jpg}}
 	\hfil
-	\subfloat[\label{fig:view2}]{\includegraphics[scale=0.10]{./images/half_artemis02.JPG}}
+	\subfloat[\label{fig:view2}]{\includegraphics[scale=0.10]{./half_artemis02.JPG}}
 	\hfil
-	\subfloat[\label{fig:mesh}] {\includegraphics[scale=0.12]{./images/half_artemis00.png}}
[…]
