### 2016-08-05 #47 @labarba
Thanks for jumping in for JOSS, @davclark !!
