### 2016-08-05 #47 @charalambos
That is correct in the case where you would like to scan a new object. For testing however, we provide a complete dataset (Alexander folder) which you can use.
