### 2016-08-05 #47 @v3c70r
The only hardware required is a nVidia graphics card supports CUDA. We used Geforce GTX770 in our development.
