### 2016-08-08 #47 @pjotrp
How large it the git repo? I interrupted at 100Mb. Usually it is a good idea to have the data separate from the sources.
