### 2016-08-08 #47 @v3c70r
@pjotrp  The repository is 299.04MB. I have no problem to clone the repository. But you are right that it is better to have the data separated.
