### 2016-08-09 #47 @pjotrp
OK, put a warning on the size in the README, that should suffice for now.
