+++ README.md
@@ -1,9 +1,13 @@
 # 3DUNDERWORLD-Structured-Light-Scanner
 =====================================
 
-[v4] Developed by: Qing Gu, Charalambos Poullis (http://www.poullis.org). Immersive and Creative Technologies Lab (http://www.theICTlab.org), Concordia University
+[v4] Lead software developer: Qing Gu
+     Researchers: Qing Gu, Charalambos Poullis
+     Immersive and Creative Technologies Lab (http://www.theICTlab.org), Concordia University
 
-[v1-v3.2] Developed by: Kyriakos Herakleous, Charalambos Poullis (http://www.poullis.org). Immersive and Creative Technologies Lab (http://www.theICTlab.org), Cyprus University of Technology
+[v1-v3.2] Lead software developer: Kyriakos Herakleous
+          Researchers: Kyriakos Herakleous, Charalambos Poullis
+          Immersive and Creative Technologies Lab (http://www.theICTlab.org), Concordia University
 
 Part of the 3DUNDERWORLD project: http://www.3dunderworld.org
 
