### 2016-08-12 79d34a3
updated bibtex for old paper
