### 2016-09-23 #47 @Kevin-Mattheus-Moerman
@simonfuhrmann @kangxue @mikhail-matrosov @daeyun @lasvegasrc @daviddoria @krm15 @kysucix is this your cup of tea? Are you willing to sign up as a [JOSS](joss.theoj.org) reviewer and review this submission?

Link to the submission in JOSS:
https://github.com/openjournals/joss-reviews/issues/47

The project GitHub repository:
https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU

The paper:
https://github.com/openjournals/joss-reviews/files/402839/10.21105.joss.00047.pdf

The submission's project website:
http://www.3dunderworld.org/

Reviewer instructions:
http://joss.theoj.org/about#reviewer_guidelines
