### 2016-09-23 #47 @Kevin-Mattheus-Moerman
@charalambos Some editorial comments on  your paper.
Could all authors share the same affiliation index? (currently affiliations are copied 3 times in paper).
In your paper you mention:

> We have performed extensive testing with a wide range of models and the results are documented in our report.

Could you expand on this please. This would be very helpful for the current reviewers but also future readers. Where are the models? What of models are these? What is the "report"? Could you even say something like "The README will guide the user to a suite test models which allow the user to replicate typical results" (adjust as you see fit)
