### 2016-09-23 #47 @Kevin-Mattheus-Moerman
@rougier do you know an additional reviewer at INRIA perhaps?
