### 2016-09-25 #47 @daviddoria
All,

Please excuse the "newbie" questions as I'm new to this process :).

I started checking boxes at the top of this issue for things that I verified. Presumably for things that I do NOT check there should be somewhere to explain why I did not check them?

I am having trouble building the project. I have listed the first two issues below - is this where I should discuss these things? Can/should I contact the author directly to get help like this and then ensure the documentation is updated?

1) This error:
/home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/src/lib/GrayCode/GrayCode.hpp:2:30: fatal error: opencv2/opencv.hpp: No such file or directory
 #include <opencv2/opencv.hpp>

seems resolvable by adding a cmake line:
INCLUDE_DIRECTORIES(${OpenCV_INCLUDE_DIRS})

2) I then get this error:

In file included from /home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/./src/lib/core/Camera.h:24:0,
                 from /home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/./src/lib/core/fileReader.h:19,
                 from /home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/./src/lib/calibration/Calibrator.hpp:2,
                 from /home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/src/lib/calibration/Calibrator.cpp:1:
/home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/./src/lib/core/Dynamic_Bitset.h:29:23: fatal error: glm/glm.hpp: No such file or directory
 #include <glm/glm.hpp>

I tried `apt-get install glm-dev` but that does not seem to be a package (Ubuntu 14.04).

Thanks,

David
