### 2016-09-25 #47 @daviddoria
@Kevin-Mattheus-Moerman FYI the link you posted in your post "is this your cup of tea? Are you willing to sign up as a JOSS reviewer and review this submission?" links to https://github.com/openjournals/joss-reviews/issues/joss.theoj.org which seems to be broken.
