### 2016-09-25 #47 @daviddoria
@charalambos The link to the paper https://github.com/openjournals/joss-reviews/files/402839/10.21105.joss.00047.pdf is an extremely short document with no images, etc. However I see in your /doc folder that there are a bunch of image files. Is there a more "real" document that I'm missing?
