### 2016-09-25 #47 @v3c70r
@daviddoria GLM can be installed with `sudo apt install libglm-dev` on Ubuntu.
