### 2016-09-25 #47 @daviddoria
I'm getting this error:
/home/doria/source/3DUNDERWORLD-SLS-GPU_CPU/src/lib/calibration/Calibrator.hpp:41: error: ‘vector’ in namespace ‘cv’ does not name a type
     static cv::vector<cv::Point2f>  manualMarkCheckBoard(cv::Mat img);
            ^
