### 2016-09-25 #47 @daviddoria
I saw this in Line 14 of Calibrator.hpp:

static std::condition_variable cv;

That seems really fragile given that there is the 'cv' namespace for OpenCV.
