### 2016-09-25 #47 @daviddoria
It seems as though an "in source" build is expected (which is almost never recommended). I say that because:
1) The readme says "Binaries are compiled and located in the `bin` folder." which presumably is implying [sourceTree]/bin.

2) When trying to run the executable (I cloned in ~/source/underworld and built in ~/build/underworld), I get:

doria@doria-home:~/build/underworld$ ./bin/SLS
 No image read from ../../data/alexander/rightCam/dataset1/ ...  No image read from ../../data/alexander/leftCam/dataset1/ ... Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Segmentation fault (core dumped)

which seems to be referencing files relative to the build tree (which is expected to be the source tree). This is bad - these paths should be specifiable on the command line.
