### 2016-09-26 #47 @Kevin-Mattheus-Moerman
@daviddoria thanks again for reviewing this submission. To answer your questions.

> I started checking boxes at the top of this issue for things that I verified. Presumably for things that I do NOT check there should be somewhere to explain why I did not check them?
> I am having trouble building the project. I have listed the first two issues below - is this where I should discuss these things? Can/should I contact the author directly to get help like this and then ensure the documentation is updated?

Yes, please post any comments and issues you want to communicate to the authors here. This way the entire review process is open and traceable.
