### 2016-09-26 #47 @Kevin-Mattheus-Moerman
@daviddoria in relation to

> @charalambos The link to the paper https://github.com/openjournals/joss-reviews/files/402839/10.21105.joss.00047.pdf is an extremely short document with no images, etc. However I see in your /doc folder that there are a bunch of image files. Is there a more "real" document that I'm missing?

Papers in JOSS are deliberately short. See some of our motivation behind it here:
http://www.arfon.org/announcing-the-journal-of-open-source-software

Here are are paper content requirements: http://joss.theoj.org/about where it says
> JOSS papers should contain the following:
> - A list of the authors of the software
> - Author affiliations
> - A short summary describing the high-level functionality of the software
> - A list of key references including a link to the software archive

and see some examples of accepted papers here:
http://joss.theoj.org/papers/popular

So the "real" document you refer to (you probably mean a classic full length paper) is not required if the software documentation is complete enough. So if you find the descriptions are lacking feel free to also demand expanded documentation.
