### 2016-09-26 #47 @Kevin-Mattheus-Moerman
@daviddoria authors are allowed to include figures in the paper however so of course you can also recommend that if you feel it is needed.
