### 2016-09-26 #47 @charalambos
@daviddoria There is also a "traditional" paper that goes along with the code and described the technique at arxiv: http://arxiv.org/abs/1406.6595

There is also a user's manual for setting up and using the software. This was written for a previous version but the setup is identical: http://www.3dunderworld.org/wp-content/uploads/2014/03/3DUNDERWORLD-SLSv3.0.pdf
