### 2016-09-26 #47 @daviddoria
@charalambos
Thanks for the link to the traditional paper. The setup instructions still do not seem to list anything about prerequisites (which version of cmake/opencv/etc is required, which packages likely need to be installed on common platforms (the GLM issue I had), etc.

Did you see my comment about the assumption about an in-source build? I was going to say that is blocking me from continuing, but suppose I could do an in-source build to continue the review for now... haha

@Kevin-Mattheus-Moerman - is it appropriate to put a link to the arXiv paper in the JOSS paper? It's really hard to start a review without knowing what the code is even supposed to be implementing.
