### 2016-09-26 #47 @v3c70r
@daviddoria
Yes, I saw your comment on the code. Sorry I didn't get chance to touch the code last weekend, I'll fix it later tonight.
