### 2016-09-26 #47 @Kevin-Mattheus-Moerman
@charalambos @daviddoria the paper may cite other and previous works. However the software submission for JOSS should stand-alone i.e. we cannot simply refer to another paper and old documentation from a previous version.

I would prefer if the authors did the following.
1) Expand the paper to give a more complete description of the software functionality. Consider including an overview figure to graphically show the software capabilities. The paper can remain short but should give readers and reviewers enough information to understand the capabilities of the software. As long as the JOSS paper provides a sufficient "stand-alone" description, feel free to also refer to previous work/papers.
3) Please fix the documentation based on the reviewers comments. Also please update the documentation for the version submitted here. If the documentation for the previous version is indeed very similar it should be little work to update.
