### 2016-10-01 3ad9532
Updated README and parameters
