+++ README.md
@@ -1,5 +1,112 @@
 # 3DUNDERWORLD-Structured-Light-Scanner
-=====================================
+[![status](http://joss.theoj.org/papers/4329bcbc7bba33961a5e749dcacb995b/status.svg)](http://joss.theoj.org/papers/4329bcbc7bba33961a5e749dcacb995b)
+
+Structured light scanner is a tool to reconstruct point cloud from series of images lit by patterned light. As showing in the following image, patterns are projected to the object in the grid manner. Each cell in the grid has a unique sequence. The corresponding points can be extracted by decoding the sequences from both images.
+![img](https://raw.githubusercontent.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU/dev/screenshots/flow.png)
+
+In this version, the reconstruction process is reimplemented both on CPU and GPU. Both versions accelorate the reconstruction time. Especially the later one. 
+
+## Dependencies
+* OpenCV2
+* CUDA (6.0+)
+* glm
+
+## Howto
+### Compile and run demo binaries
+Demo binaries and data is included.  A CMake script is included to compile the binaries and libraries. 
+```bash
+git clone https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU.git
+cd 3DUNDERWORLD-SLS-GPU_CPU
+mkdir build
+cd build
+cmake ..
+make
+```
+If CUDA is detected in your system, both CPU and GPU constructors are created in the `bin` folder. Other wise, only CPU constructor will be generated. To run the binaries, you can pass the data folder and camera configuration files as parameters, or using default value if you compiled with the commands above.
+```
+  SLS/SLS_GPU [OPTION...]
+
+  -l, --leftcam arg      Left camera image folder (default:
+                         ../../data/alexander/leftCam/dataset1/)
+  -L, --leftconfig arg   Left camera configuration file (default:
+                         ../../data/alexander/leftCam/calib/output/calib.xml)
+  -r, --rightcam arg     Right camera image folder (default:
+                         ../../data/alexander/rightCam/dataset1/)
+  -R, --rightconfig arg  Right camera configuration file (default:
+                         ../../data/alexander/rightCam/calib/output/calib.xml)
+  -o, --output arg       Output file (default: test.ply)
+  -h, --help             Print this help
+```
+The configuration files are `xml` files generated by OpenCV calibration tool, including intrinsic and extrinsic parameters. 
+
+### Using the library
+
+This project also privides libraries that can be easily integrated into other projects. Here's a quick start code for using the libraries.
+
+```C++
+#ifdef CPU
+
+/* Using CPU reconstruction */
+#include <core/fileReader.h>
+#include <core/Reconstructor.h>
+#else
+
+/* Using CUDA reconstruction */
+#include <ReconstructorCUDA/fileReaderCUDA.cuh>
+#include <ReconstructorCUDA/ReconstructorCUDA.cuh>
+#endif
+
+
+int main()
+{
+    std::string rightCameraFolder = "../../data/alexander/rightCam/dataset1/"
+    std::string leftCameraFolder = "../../data/alexander/leftCam/dataset1/"
+    std::string rightConfigFile = "../../data/alexander/rightCam/calib/output/calib.xml"
+    std::string leftConfigFile = "../../data/alexander/leftCam/calib/output/calib.xml"
+
+    // Allocate two cameras from heap. Parameters are the names of cameras. 
+#ifdef CPU
+    SLS::FileReader *rightCam=new SLS::FileReader("rightCamera");
+    SLS::FileReader *leftCam= new SLS::FileReader("leftCamera");
+#else
[…]
