+++ README.md
@@ -4,7 +4,7 @@
 Structured light scanner is a tool to reconstruct point cloud from series of images lit by patterned light. As showing in the following image, patterns are projected to the object in the grid manner. Each cell in the grid has a unique sequence. The corresponding points can be extracted by decoding the sequences from both images.
 ![img](https://raw.githubusercontent.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU/dev/screenshots/flow.png)
 
-In this version, the reconstruction process is reimplemented both on CPU and GPU. Both versions accelorate the reconstruction time. Especially the later one. 
+In this version, the reconstruction process is reimplemented both on CPU and GPU. Both versions accelerate the reconstruction time. Especially the later one. 
 
 ## Dependencies
 * OpenCV2
@@ -22,7 +22,7 @@ cd build
 cmake ..
 make
 ```
-If CUDA is detected in your system, both CPU and GPU constructors are created in the `bin` folder. Other wise, only CPU constructor will be generated. To run the binaries, you can pass the data folder and camera configuration files as parameters, or using default value if you compiled with the commands above.
+If CUDA is detected in your system, both CPU and GPU constructors are created in the `bin` folder. Otherwise, only CPU constructor will be generated. To run the binaries, you can pass the data folder and camera configuration files as parameters, or using the default value if you compiled with the commands above.
 ```
   SLS/SLS_GPU [OPTION...]
 
@@ -41,7 +41,7 @@ The configuration files are `xml` files generated by OpenCV calibration tool, in
 
 ### Using the library
 
-This project also privides libraries that can be easily integrated into other projects. Here's a quick start code for using the libraries.
+This project also provides libraries that can be easily integrated into other projects. Here's a quick start code for using the libraries.
 
 ```C++
 #ifdef CPU
