+++ README.md
@@ -39,7 +39,7 @@ If CUDA is detected in your system, both CPU and GPU constructors are created in
 ```
 The configuration files are `xml` files generated by OpenCV calibration tool, including intrinsic and extrinsic parameters. 
 
-### Using the library
+### Use the library
 
 This project also provides libraries that can be easily integrated into other projects. Here's a quick start code for using the libraries.
 
@@ -106,7 +106,7 @@ We welcome you to implement your camera class and make a pull request to this pr
 
 Also please note that since the example data files are currently included in the repository, the size of repository is 300MB. It may take while to clone. 
 
-## Version Informations
+## Version Information
 
 ### [v4] 
 Lead software developer: Qing Gu
