### 2016-10-06 #47 @Kevin-Mattheus-Moerman
@charalambos @v3c70r any progress on addressing the comments by myself and the reviewer?

@daviddoria thanks again for reviewing this submission. Could you update me on whether you are currently continuing actively reviewing content or if you are waiting for action by the authors?  Thanks.
