### 2016-10-06 #47 @v3c70r
@daviddoria @Kevin-Mattheus-Moerman  Yes.
1)  Now data and configuration files are passed as parameters. The default of the parameters can be used when the bin is compiled in `[sourceTree]/bin`.
2) The readme documents has been updated to contain more information, including the purpose of the tool, an instructions of the sample code and an example of using the library.
