### 2016-10-09 #47 @daviddoria
@v3c70r Thanks, it seems to build now, and the summary in the readme is much much better. I tried to produce the help message, but I get:

doria@doria-home:~/build/underworld$ ./bin/SLS -h
terminate called after throwing an instance of 'std::regex_error'
  what():  regex_error
Aborted (core dumped)
