### 2016-10-09 #47 @daviddoria
@v3c70r Actually, I'm getting that no matter what I try:

doria@doria-home:~/build/underworld$ ./bin/SLS
terminate called after throwing an instance of 'std::regex_error'
  what():  regex_error
Aborted (core dumped)

doria@doria-home:~/build/underworld$ ./bin/SLS -l ~/source/3DUNDERWORLD-SLS-GPU_CPU/data/alexander/leftCam/dataset1/ -L ~/source/3DUNDERWORLD-SLS-GPU_CPU/data/alexander/leftCam/calib/output/calib.xml -r ~/source/3DUNDERWORLD-SLS-GPU_CPU/data/alexander/rightCam/dataset1/ -R ~/source/3DUNDERWORLD-SLS-GPU_CPU/data/alexander/rightCam/calib/output/calib.xml -o test.ply
terminate called after throwing an instance of 'std::regex_error'
  what():  regex_error
Aborted (core dumped)
