### 2016-10-09 #47 @v3c70r
@daviddoria Hmm, it seems no problem on my machine. What compiler are you using?
