### 2016-10-09 #47 @daviddoria
@v3c70r I commented everything in main() and the error was still there. I commented the headers one at a time and as soon as I stop including cppopts.hpp the error goes away. Might I recommend using boost::program_options instead of 1400 lines of custom parsing? (see my compiler and system in my previous comment)
