### 2016-10-09 #47 @v3c70r
@daviddoria Right, I'm removing that header and use `cmdline.h` instead.
