### 2016-10-09 289c64e
Using cmdline.h arg parser for both GPU and CPU reconstructors
