+++ src/app/App.cpp
@@ -3,34 +3,25 @@
 #include <core/log.hpp>
 #include <glm/gtx/string_cast.hpp>
 #include <core/ReconstructorCPU.h>
-#include "cxxopts.hpp"
+#include "cmdline.h"
 
 int main(int argc, char** argv)
 {
     
-    cxxopts::Options options("SLS_CPU", "CPU implementation of SLS");
-    options.add_options()
-        ("l,leftcam", "Left camera image folder", cxxopts::value<std::string>()->default_value("../../data/alexander/leftCam/dataset1/"))
-        ("L,leftconfig", "Left camera configuration file", cxxopts::value<std::string>()->default_value("../../data/alexander/leftCam/calib/output/calib.xml"))
-        ("r,rightcam", "Right camera image folder", cxxopts::value<std::string>()->default_value("../../data/alexander/rightCam/dataset1/"))
-        ("R,rightconfig", "Right camera configuration file", cxxopts::value<std::string>()->default_value("../../data/alexander/rightCam/calib/output/calib.xml"))
-        ("o,output", "Output file", cxxopts::value<std::string>()->default_value("test.ply"))
-        ("h,help","Print this help")
-        ;
-    options.parse(argc, argv);
-
-    if (options.count("help"))
-    {
-        std::cout<<options.help()<<std::endl;
-        return 0;
-    }
+    cmdline::parser p;
+    p.add<std::string>("leftcam", 'l',"Left camera image folder", false, "../../data/alexander/leftCam/dataset1/");
+    p.add<std::string>("rightcam", 'r',"Right camera image folder", false, "../../data/alexander/rightCam/dataset1/");
+    p.add<std::string>("leftconfig", 'L',"Left camera configuration file", false, "../../data/alexander/leftCam/calib/output/calib.xml");
+    p.add<std::string>("rightconfig", 'R',"Right camera configuration file", false, "../../data/alexander/rightCam/calib/output/calib.xml");
+    p.add<std::string>("output", 'o',"Right camera configuration file", false, "output.ply");
+    p.parse_check(argc, argv);
 
     LOG::restartLog();
-    std::string leftCameraFolder = options["l"].as<std::string>();
-    std::string rightCameraFolder = options["r"].as<std::string>();
-    std::string leftConfigFile = options["L"].as<std::string>();
-    std::string rightConfigFile = options["R"].as<std::string>();
-    std::string output = options["output"].as<std::string>();
+    std::string leftCameraFolder = p.get<std::string>("leftcam");
+    std::string rightCameraFolder = p.get<std::string>("rightcam");
+    std::string leftConfigFile = p.get<std::string>("leftconfig");
+    std::string rightConfigFile = p.get<std::string>("rightconfig");
+    std::string output = p.get<std::string>("output");
 
     SLS::FileReader *rightCam=new SLS::FileReader("rightCamera");
     SLS::FileReader *leftCam= new SLS::FileReader("leftCamera");
