### 2016-10-09 #47 @v3c70r
@daviddoria Updated, the command line parser should work now.
