+++ README.md
@@ -39,6 +39,9 @@ If CUDA is detected in your system, both CPU and GPU constructors are created in
 ```
 The configuration files are `xml` files generated by OpenCV calibration tool, including intrinsic and extrinsic parameters. 
 
+### Grab the output
+Most of the outputs are written to the log file named `SLS.log`. To track the outputs while running the binary, run `tail -f sls.log` in another terminal of the same directory. 
+
 ### Use the library
 
 This project also provides libraries that can be easily integrated into other projects. Here's a quick start code for using the libraries.
