### 2016-10-09 #47 @daviddoria
@v3c70r That worked. Next comment - it finally built and started to run, but I got:
--------------
Starting /home/doria/build/underworld/bin/SLS...
OpenCV Error: Unspecified error (The function is not implemented. Rebuild the library with Windows, GTK+ 2.x or Carbon support. If you are on Ubuntu or Debian, install libgtk2.0-dev and pkg-config, then re-run cmake or configure script) in cvWaitKey, file /home/doria/source/opencv-2.4.13/modules/highgui/src/window.cpp, line 567
terminate called after throwing an instance of 'cv::Exception'
  what():  /home/doria/source/opencv-2.4.13/modules/highgui/src/window.cpp:567: error: (-2) The function is not implemented. Rebuild the library with Windows, GTK+ 2.x or Carbon support. If you are on Ubuntu or Debian, install libgtk2.0-dev and pkg-config, then re-run cmake or configure script in function cvWaitKey

The program has unexpectedly finished.
--------------

Luckily, the error told me exactly how to fix it. However, this should be caught at configure time, not run time (surely there's a CMake query to ask OpenCV if it has whatever GTK stuff you require).

It did seem to run to completion now, so I can continue the review!
