### 2016-10-09 #47 @daviddoria
I tried to view the output .ply file in Paraview (5.1) and got:

ERROR: In /home/kitware/dashboards/buildbot/paraview-debian6dash-linux-shared-release_opengl2_qt4_superbuild/source-paraview/VTK/IO/PLY/vtkPLYReader.cxx, line 138
vtkPLYReader (0x31a7780): Cannot read geometry

It did open OK in Meshlab 1.3.2 - but I really doubt it's a bug in VTK's PLY reader, so it might be worth looking into if something is non-standard in your file (and Meshlab is "accidentally" reading it correctly?)
