### 2016-10-09 #47 @v3c70r
@daviddoria Thank you for your input. The ply file works fine with Blender.
, I will investigate the mesh exporter and do some more tests.
