### 2016-10-09 #47 @daviddoria
@v3c70r Please find some comments below. Recall this is my first review, so if these types of comments are not what we're looking for, then someone please say so :)

- There are several key functions (like Calibrate(...) in Calibrator.hpp that do not have descriptions (the class itself should also have a description).

- All of the App*.cpp files need descriptions (and probably better names - perhaps using action verbs? e.g. App_Calib.cpp -> CalibrateImages.cpp, App.cpp -> ComputeReconstruction.cpp).

- The classes in the core/ folder are not well commented. E.g. (but this really applies to all of them) Camera.h and fileReader.h have tons of functions with no description. They are definitely not "obvious" either - I am left wondering why the Camera class has a mask, what a white/black threshold is, what kind of distortion 'undistort()' is correcting, what kind of configFile it can read, etc. I also am wondering why the fileReader class has getRay accessors, what kind of files it reads, why it seems to duplicate the Camera classes "white threshold" accessor, etc.

- (minor) - there doesn't seem to be a convention for file naming. There is log.hpp (all lower case), Camera.h (first character upper case), Dynamic_Bitset.h (underscore between words), fileReader.h (no underscore between words), etc.

- From looking at Reconstructor.h, I don't understand the API. What does it mean to add a camera to this object? It's not clear that pointCloud_ (I think) is an output object. I don't know what a Projector is. what is idx_? Why does reconstruct() not return the point cloud? etc.

- The GrayCode class has zero comments. Why is it in its own folder and not just in core/?

- (minor) a few files (Ray.h, Projector.h) do not show up in my QtCreator project. Simply listing them in core/CMakeLists.txt like this will fix that:
add_library(core ${SOURCES} Ray.h Projector.h)

- CRITICAL: there do not seem to be any unit tests at all. This is very very bad. Since you include the nice Alexander data set, at the very least you could execute the reconstruction and compare the points that are produced to a "known-good" baseline reconstruction.

--

My overall feeling is that this code would be very very hard for someone to extend. I would like the core reconstruction algorithm to be heavily documented (if ReconstructorCPU::reconstruct() is even the right place to be looking, there are nearly zero comments).
