### 2016-10-13 #47 @mikhail-matrosov
Review on commit d9786db.

Greetings, everyone!

I tried to build the code, here's what i faced:
1. cmake didn't check if libglm-dev was installed.
2. make command failed probably because of wrong GCC options (btw, i have Ubuntu 15.10):
```
make
[  6%] Building NVCC (Device) object src/lib/ReconstructorCUDA/CMakeFiles/sls_gpu.dir/sls_gpu_generated_ReconstructorCUDA.cu.o
/usr/lib/gcc/x86_64-linux-gnu/4.9/include/stddef.h(432): error: identifier "nullptr" is undefined
/usr/lib/gcc/x86_64-linux-gnu/4.9/include/stddef.h(432): error: expected a ";"
...<another hundred lines of errors>...
```
I also tried GCC 5.2.1 and 4.7.4 with the same result. Suggest that cmake should at least check compiler's version. A much better solution would be to fix the code so it can be build with GCC 5.

Also it would be great if it was possible to build the code without CUDA (even if it was detected in the system).
