### 2016-10-13 #47 @vaheta
I've tried to compile the code, couldn't yet do it. Getting the same error as @mikhail-matrosov. I'll try later on another PC without CUDA, meanwhile here are my thoughts on Reviewer questions:

Conflict of interest

**YES** As the reviewer I confirm that there are no conflicts of interest for me to review this work (such as being a major contributor to the software).

General checks

**YES** Repository: Is the source code for this software available at the repository url?
**YES** License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
**YES** Version: Does the release version given match the GitHub release (v4.0.0)?

Functionality

**NO** Installation: Does installation proceed as outlined in the documentation? - getting errors, couldn't fix them so far. Also, as mentioned above, cmake doesn't check if glm is installed.
**-** Functionality: Have the functional claims of the software been confirmed? - will be able to check this after error fixed
**-** Performance: Have any performance claims of the software been confirmed? - will be able to check this after error fixed

Documentation
I've written this based on this: http://www.3dunderworld.org/wp-content/uploads/2014/03/3DUNDERWORLD-SLSv3.0.pdf and this: https://arxiv.org/pdf/1406.6595v2.pdf. Although, it is hard to call the paper "documentation", would be nice to have a separate documentation file. Also, the README.md file has unique information compared to previous two files. Again, it would be better to have a single full documentation file with detailed description on how to install, run and extend the software with some examples, and README.md should be just a short version of that doc.
**YES** A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
**NO** Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. - couldn't find them in any of these two files.
**PARTLY** Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems). - There is an explanation on how to use in README.md file and different explanation in documentation file. No single place with detailed description.
**PARTLY** Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)? - One can find most information about functionality after reading the paper, docs and README.md, but it will be much better to present everything in a more structured way in a single doc.
**-** Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified? -  - will be able to check this after error fixed
**NO/YES/YES** Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support - it is hard to contribute without structured description of all the parts of the software and without extensively commented code. For 2) and 3) I guess Github's "Issues" section is good enough.

Software paper

Paper PDF: 10.21105.joss.00047.pdf

**YES** Authors: Does the paper.md file include a list of authors with their affiliations?
**YES** A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
**NO** References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)? - couldn't find the link to original paper.

Overall, I think that making the software more robust to different configurations and better structured documentation are main things to do for now. Developing good Structured Light software is a good goal, and, when it will be ready and well documented, it will be quiet useful for the community.
Also, it may be a good idea to add photo capture process through something like gphoto2 for arbitrary camera.
