### 2016-10-14 #47 @v3c70r
@vaheta @davclark Thank you so much for your review, they are very helpful. I'll update the code accordingly this weekend.
@vaheta @mikhail-matrosov  It looks like the nvcc c++11 is not explicitly enabled in cmake script. I will fix it.

>Also it would be great if it was possible to build the code without CUDA (even if it was detected in the system).

It's a good idea, I'll add the option.
