### 2016-10-21 a8761b7
Added ENABLE_CUDA flag in CMake, added -std=c++11 flag to NVCC
