+++ src/app/CMakeLists.txt
@@ -2,8 +2,7 @@ aux_source_directory(. SOURCES)
 add_executable(SLS App.cpp)
 target_link_libraries( SLS core)
 
-IF(${CUDA_FOUND})
-    list(APPEND CUDA_NVCC_FLAGS "-D__STRICT_ANSI__ -D_MWAITXINTRIN_H_INCLUDED -D_FORCE_INLINES")
+IF(${USE_CUDA})
     cuda_add_executable(SLS_GPU ./App_CUDA.cu)
     target_link_libraries(SLS_GPU core sls_gpu)
     set_target_properties ( SLS_GPU
@@ -12,7 +11,7 @@ IF(${CUDA_FOUND})
         #LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib"
         RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin"
         )
-ENDIF(${CUDA_FOUND})
+ENDIF(${USE_CUDA})
 
 add_executable(SLS_CALIB App_Calib.cpp)
 target_link_libraries(SLS_CALIB core sls_calib)
