### 2016-10-24 5ac0b19
Removed dead code, refactoring expoter
