+++ refCode/RNG.h
@@ -1,291 +1,0 @@
-//------------------------------------------------------------------------------------------------------------
-//* Copyright © 2010-2015 Immersive and Creative Technologies Lab, Cyprus University of Technology           *
-//* Link: http://www.theICTlab.org                                                                           *
-//* Software developer(s): Kyriakos Herakleous                                                               *
-//* Researcher(s): Kyriakos Herakleous, Charalambos Poullis                                                  *
-//*                                                                                                          *
-//* License: Check the file License.md                                                                       *
-//------------------------------------------------------------------------------------------------------------
-
-
-//random number generator
-#ifndef __RNG_H__
-#define __RNG_H__
-
-#include <math.h>
-#include <stdlib.h>
-
-#define IA   16807
-#define IM   2147483647
-#define IQ   127773
-#define IR   2836
-#define MASK 123459876
-#define NTAB 32
-#define EPS  1.2e-7
-#define RNMX (1.0-EPS)
-
-
-class RNG
-{
-public:
-   RNG(unsigned long long _seed = 7564231ULL) 
-   {
-      seed      = _seed;
-      mult      = 62089911ULL;
-      llong_max = 4294967295ULL;
-      double_max = 4294967295.0f;
-   }
-   double operator()();
-	
-   double uniformRandCLib(double _min,double _max);
-   double fastUniformRand(double _min, double _max);
-   double uniformRand(double _min, double _max);
-   double normalRand(void);
-   double normalRand(double mu, double sigma);
-
-
-   unsigned long long seed;
-   unsigned long long mult;
-   unsigned long long llong_max;
-   double double_max;
-};
-
-inline double RNG::operator()()
-{
-   seed = mult * seed;
-   return double(seed % llong_max) / double_max;
-}
-
-/*! ************************************************************************* 
-
-    Function:
-        CRandomNumberGenerator::FastUniformRand()
-
-    Description:
-        Generates a random number between min and max inclusively using a
-        uniform random deviate.
-
-    Parameters:
-        min - the minimum number that can be generated
-        max - the maximum number that can be generated
-
-    Return Value:
-        A random number between min and max inclusively.
-
-    Notes:
-        This algoirthm is adapted from:
-
-            Press, W, Teukolsky, S, et al., "Numerical Recipes in C", 2nd Ed,
-            pp 283-285, Cambridge Press, 1996.
-
-*****************************************************************************/
-
-
-inline double RNG::fastUniformRand(double _min, double _max) {
-
-    int k;
-    double ans;
-
-    seed ^= MASK;                     // XORing with Mask allows use of zero for idnum
-
-    k = seed / IQ;
-
-    seed = IA*(seed-k*IQ) - IR*k;   // compute idnum=(IA*idum) % IM iwth overflows by Schrage's method
-    if( seed < 0) {
-        seed += IM;
-    }
-
-    ans = (1.0f/IM)*seed;             // convert idum to a doubleing result
-
-    seed ^= MASK;                     // unmask before return
-
-    ans = ans*(_max-_min) + _min;            // shift number into desired range
-
-    return ans;
-}
-
-
-/*! ************************************************************************* 
-
-    Function:
-        CRandomNumberGenerator::UniformRandCLib()
-
-    Description:
-        Generates a random number between min and max inclusively using the
-        uniform random deviate from the C standard library.  It is beter to
-        use the function UniformRand().  For a discussion see:
-
-            Press, W, Teukolsky, S, et al., "Numerical Recipes in C", 2nd Ed,
-            pp 275-277, Cambridge Press, 1996.
-
-    Parameters:
-        min - the minimum number that can be generated
[…]
