### 2016-10-27 1e32c24
CUDA reconstruction reconstructed grid image with random color (garbage value) assigned.
