+++ src/lib/ReconstructorCUDA/ReconstructorCUDA.cu
@@ -35,11 +35,13 @@ void ReconstructorCUDA::reconstruct()
     /**/
 
 
+    // For each camera
     for(size_t camIdx = 0; camIdx < cameras_.size(); camIdx++)
     {
+        // Cast camera to CUDA camera
         FileReaderCUDA* cam = (FileReaderCUDA*)cameras_[camIdx];
         LOG::writeLog("Generating reconstruction bucket for \"%s\" ... \n", cam->getName().c_str());
-        cam->computeShadowsAndThresholds();    // can it be done in GPU? Yes
+        cam->computeShadowsAndThresholds();    // TODO: Put this part to GPU too.
         size_t x=0,y=0,xTimesY=0;
         cam->getResolution(x,y);
         xTimesY=x*y;
@@ -48,10 +50,16 @@ void ReconstructorCUDA::reconstruct()
         cam->getNextFrame();
         // Load all images into GPU memory
         uchar *images_d=nullptr;
+
+        // Allocate image memory on GPU with (number of pixel) * (number of pattern image + number of reversed pattern image)
         gpuErrchk(cudaMalloc((void**)&images_d, sizeof(uchar)*xTimesY*projector_->getRequiredNumFrames()*2));
+
+        // Initialize binary sequence for each pixel. Here we have xTimesY pixels and 
+        // each pixel has `projector_->getRequiredNumFrames()` bits.
         Dynamic_Bitset_Array bitsetArray(xTimesY, projector_->getRequiredNumFrames());
 
-        //Preparing data
+        // Preparing data
+        // Copy images to GPU
         for (size_t i=0; i<projector_->getRequiredNumFrames(); i++)
         {
             auto frm = cam->getNextFrame();
@@ -63,6 +71,7 @@ void ReconstructorCUDA::reconstruct()
                     sizeof(uchar)*xTimesY, cudaMemcpyHostToDevice));
         }
 
+        // Generate bit array for all pixels from image sequence.
         Kernel::genPatternArray<<<200,200>>> (
                 images_d, 
                 projector_->getRequiredNumFrames(),
@@ -71,10 +80,11 @@ void ReconstructorCUDA::reconstruct()
                 cam->getMask()->getGPUOBJ(),
                 bitsetArray.getGPUOBJ()
                 );
-        //Check for errors
+        // Check for errors
         gpuErrchk(cudaPeekAtLastError());
         gpuErrchk(cudaFree(images_d)); // Release the heavy image array
 
+        // Insert pixels into bucket
         Kernel::buildBuckets<<<200, 200>>> (
              cam->getMask()->getGPUOBJ(),
              bitsetArray.getGPUOBJ(),
@@ -90,11 +100,11 @@ void ReconstructorCUDA::reconstruct()
     // TODO: Export point cloud in (x, y, z, r, g, b) format
     auto camera0 = (FileReaderCUDA*)(cameras_[0]);
     auto camera1 = (FileReaderCUDA*)(cameras_[1]);
-    float* cloud = nullptr;
+    float* pointCloud_d = nullptr;  // Point cloud on device with x,y,z,r,g,b.
     size_t resX, resY;
     camera0->getResolution(resX, resY);
 
-    gpuErrchk ( cudaMalloc((void**)&cloud, buckets[0].getNumBKTs()*sizeof(float)*6));
+    gpuErrchk ( cudaMalloc((void**)&pointCloud_d, buckets[0].getNumBKTs()*sizeof(float)*6));
 
     // Reconstructing point cloud
     LOG::writeLog("Reconstructing point cloud ...\n");
[…]
