### 2016-10-31 e150183
Mesh exporter refactored, but color doesn't look right
