+++ src/lib/core/exporter.cpp
@@ -1,5 +1,6 @@
 #include "exporter.hpp"
 #include <fstream>
+#include <sstream>
 
 namespace SLS
 {
@@ -47,64 +48,6 @@ void exportPLYGrid(std::string fileName, const Reconstructor& reconstructor)
             of<<"0"<<std::endl;
 }
 
-void exportPLY( std::string fileName ,const Reconstructor& reconstructor)
-{
-    LOG::startTimer("Exporting PLY to %s ... ", fileName.c_str());
-
-    std::ofstream of(fileName, std::ofstream::out);
-    const auto &pointCloud = reconstructor.pointCloud_;
-    // Writing Headers
-    of <<"ply\n"<<
-        "format ascii 1.0\n"<<
-        "element vertex "<<pointCloud.size()/6<<std::endl<<
-        "property float x\n"<<
-        "property float y\n"<<
-        "property float z\n"<<
-        "property float red\n"<<
-        "property float geen\n"<<
-        "property float blue\n"<<
-        //"property list uchar int vertex_indices\n"<<
-        "end_header\n";
-    // Writing vertex list
-    // Writing face list
-    for (size_t i=0; i<pointCloud.size(); i+=6)
-        of<<pointCloud[i+0]<<" "<<pointCloud[i+1]<<" "<<pointCloud[i+2]<<" "
-        <<pointCloud[i+3]<<" "<<pointCloud[i+4]<<" "<<pointCloud[i+5]<<std::endl;
-    of.close();
-
-    LOG::endTimer();
-}
-
-void exportOBJ( std::string fileName ,const Reconstructor& reconstructor)
-{
-    LOG::startTimer("Exporting OBJ to %s ... ", fileName.c_str());
-
-    std::ofstream of(fileName, std::ofstream::out);
-    const auto &pointCloud = reconstructor.pointCloud_;
-    for (size_t i=0; i<pointCloud.size(); i+=6)
-        of<<"v "<<pointCloud[i]<<" "<<pointCloud[i+1]<<" "<<pointCloud[i+2]<<std::endl;
-    of.close();
-
-    LOG::endTimer();
-}
-
-void exportOBJVec4( std::string fileName ,const Reconstructor& reconstructor)
-{
-    LOG::startTimer("Exporting OBJ vec4 to %s ... ", fileName.c_str());
-
-    std::ofstream of(fileName, std::ofstream::out);
-    const auto &pointCloud = reconstructor.pointCloud_;
-    for (size_t i=0; i<pointCloud.size(); i+=4)
-    {
-        if (pointCloud[i+3] < 0.5 ) continue;
-        of<<"v "<<pointCloud[i]<<" "<<pointCloud[i+1]<<" "<<pointCloud[i+2]<<std::endl;
-    }
-    of.close();
-
-    LOG::endTimer();
-}
-
-
 // Refactoring export functions
 
 void exportPointCloud(std::string fileName, std::string type, const Reconstructor& reconstructor)
@@ -140,10 +83,27 @@ void exportPointCloud2PLY(std::string fileName, const std::vector<float> &pointC
 {
     LOG::startTimer("Exporting PLY to %s ... ", fileName.c_str());
     std::ofstream of(fileName, std::ofstream::out);
-    // Writing Headers
-    of <<"ply\n"<<
+    std::stringstream headerSS;
+    std::stringstream dataSS;
+  
+    size_t pointCount = 0;
+    for (size_t i=0; i<pointCloud.size(); i+=6)
+    {
+        std::array<float, 6> p;
+        memcpy ( p.data(), &pointCloud[i], sizeof(float) * 6);
+        float sum = 0.0;
[…]
