+++ src/app/App_Graycode.cpp
@@ -1,8 +1,0 @@
-#include <iostream>
-#include <GrayCode/GrayCode.hpp>
-int main()
-{
-    SLS::GrayCode gc(1024, 768);
-    gc.generateGrayCode();
-    return 0;
-}
