+++ src/lib/core/Exporter.cpp
@@ -1,4 +1,4 @@
-#include "exporter.hpp"
+#include "Exporter.hpp"
 #include <fstream>
 #include <sstream>
 
