+++ src/app/App_CUDA.cu
@@ -1,4 +1,3 @@
-#include <ReconstructorCUDA/Dynamic_bits.cuh>
 #include <ReconstructorCUDA/FileReaderCUDA.cuh>
 #include <ReconstructorCUDA/ReconstructorCUDA.cuh>
 #include "cmdline.h"
