### 2016-11-03 6056228
Fixed num of frame calculation
