+++ src/lib/core/FileReader.cpp
@@ -5,7 +5,7 @@
 
 namespace SLS
 {
-void FileReader::loadImages(const std::string& folder, bool isGL)
+void FileReader::loadImages(const std::string& folder, std::string suffix, bool isGL)
 {
     LOG::startTimer("Loading image from %s ... ", folder.c_str());
     std::stringstream ss;
@@ -16,7 +16,8 @@ void FileReader::loadImages(const std::string& folder, bool isGL)
     while(true)
     {
         std::stringstream jpgss;
-        jpgss<<std::setfill('0')<<std::setw(4)<<images_.size()<<".jpg";
+        // TODO: fix Quick hack
+        jpgss<<std::setfill('0')<<std::setw(4)<<images_.size()<<"."<<suffix;
         std::string fName = ss.str()+jpgss.str();
         cv::Mat img=cv::imread(fName, CV_LOAD_IMAGE_COLOR);
         if (!img.data)
