### 2016-11-04 #47 @Kevin-Mattheus-Moerman
@v3c70r can you give us an update on progress? Are you able to address the issues raised above?
