### 2016-11-04 #47 @v3c70r
Yes, I'm currently working on the issues and have addressed some of them. But I need more time to finish all of them since I'm currently have a full-time job and work on the project in free time.

Here is the update:

- [x]  NVCC flags have been defined to support c++11 by default. CMake option `ENABLE_CUDA` is used turn on/off CUDA.
- [x]  Point cloud exporter should work properly now.
- [x]  Naming convention on files.

And here are things on the TODO list:

- [ ] User has reported bugs while using different projector resolutions. This issue has currently addressed in CPU implementation, but not yet in GPU.
- [ ] Unit tests.
- [ ] Documentation on the key functions and APIs.
- [ ] Refactoring, need more sensible camera and reconstructor structure.
- [ ] Update readme accordingly.
