### 2016-11-26 d12d580
Added suffix option
