### 2016-11-27 0b43889
Changed camera pointer ownership, initialized google test
