+++ test/App.cpp
@@ -3,8 +3,7 @@
 #include <core/Log.hpp>
 #include <glm/gtx/string_cast.hpp>
 #include <core/ReconstructorCPU.h>
-#include <memory>
-#include "cmdline.h"
+#include <cmdline.h>
 
 int main(int argc, char** argv)
 {
@@ -32,8 +31,8 @@ int main(int argc, char** argv)
     std::string output = p.get<std::string>("output");
     std::string suffix = p.get<std::string>("format");
 
-    auto rightCam = std::unique_ptr<SLS::FileReader>(new SLS::FileReader("rightCamera"));
-    auto leftCam = std::unique_ptr<SLS::FileReader>(new SLS::FileReader("leftCamera"));
+    SLS::FileReader *rightCam=new SLS::FileReader("rightCamera");
+    SLS::FileReader *leftCam= new SLS::FileReader("leftCamera");
 
     rightCam->loadImages(rightCameraFolder, suffix);
     leftCam->loadImages(leftCameraFolder, suffix);
@@ -42,8 +41,8 @@ int main(int argc, char** argv)
     leftCam->loadConfig(leftConfigFile);
 
     SLS::ReconstructorCPU reconstruct(p.get<size_t>("width"), p.get<size_t>("height"));
-    reconstruct.addCamera(rightCam.get());
-    reconstruct.addCamera(leftCam.get());
+    reconstruct.addCamera(rightCam);
+    reconstruct.addCamera(leftCam);
 
     reconstruct.reconstruct();
 
