+++ test/test.cpp
@@ -1,2 +1,0 @@
-// Unit test with Google Test framework
-
