+++ CMakeLists.txt
@@ -72,7 +72,7 @@ add_subdirectory(./src/lib/GrayCode)
 add_subdirectory(./src/lib/calibration)
 add_subdirectory(./src/lib/core)
 add_subdirectory(./src/app)
-#add_subdirectory(./test)
+add_subdirectory(./test)
 
 set_target_properties ( SLS
     PROPERTIES
