### 2016-12-10 f2dc516
integrated test script
