+++ src/lib/core/Exporter.cpp
@@ -2,6 +2,7 @@
 #include <fstream>
 #include <sstream>
 #include <unordered_map>
+#include <functional>
 
 namespace SLS
 {
