+++ CMakeLists.txt
@@ -19,7 +19,7 @@ if(NOT WIN32)
     set(BoldWhite   "${Esc}[1;37m")
 endif()
 
-cmake_minimum_required(VERSION 2.8)
+cmake_minimum_required(VERSION 3.1)
 set( CMAKE_EXPORT_COMPILE_COMMANDS 1 )
 project( SLS )
 
@@ -36,25 +36,28 @@ include_directories(${OpenCV_INCLUDE_DIRS})
 #Detecting CUDA
 find_package(CUDA)
 # Set default ENABLE_CUDA value
-if (NOT DEFINED ${ENABLE_CUDA})
-    if (${CUDA_FOUND})
-        set(ENABLE_CUDA on)
-    else (${CUDA_FOUND})
-        set(ENABLE_CUDA off)
-    endif(${CUDA_FOUND})
-endif (NOT DEFINED ${ENABLE_CUDA})
+option (ENABLE_CUDA "Enable CUDA in build" OFF)
+option (GTEST "Enable Google test in buid" OFF)
+#if (NOT DEFINED ${ENABLE_CUDA})
+#    if (${CUDA_FOUND})
+#        set(ENABLE_CUDA ON)
+#    else (${CUDA_FOUND})
+#        set(ENABLE_CUDA OFF)
+#    endif(${CUDA_FOUND})
+#    message( "CUDA not defined")
+#endif (NOT DEFINED ${ENABLE_CUDA})
 
 if (${CUDA_FOUND} AND ${ENABLE_CUDA})
-    set( USE_CUDA on)
+    set( USE_CUDA ON)
     include_directories(${CUDA_INCLUDE_DIRS})
     set(CUDA_PROPAGATE_HOST_FLAGS OFF)
     list(APPEND CUDA_NVCC_FLAGS "-O2 -std=c++11 -D__STRICT_ANSI__ -D_MWAITXINTRIN_H_INCLUDED -D_FORCE_INLINES")
 else (${CUDA_FOUND} AND ${ENABLE_CUDA})
-    set( USE_CUDA off)
+    set( USE_CUDA OFF)
 endif (${CUDA_FOUND} AND ${ENABLE_CUDA})
 
 if (${CUDA_FOUND} AND NOT ${ENABLE_CUDA})
-    message(STATUS "${Yellow}CUDA is available but disabled, use -DENABLE_CUDA to enable CUDA${ColourReset}") 
+    message(STATUS "${Yellow}CUDA is available but disabled, use -DENABLE_CUDA=ON to enable CUDA${ColourReset}") 
 endif (${CUDA_FOUND} AND NOT ${ENABLE_CUDA})
 
 if (NOT ${CUDA_FOUND} AND ${ENABLE_CUDA})
@@ -72,7 +75,11 @@ add_subdirectory(./src/lib/GrayCode)
 add_subdirectory(./src/lib/calibration)
 add_subdirectory(./src/lib/core)
 add_subdirectory(./src/app)
-add_subdirectory(./test)
+
+if (GTEST)
+    enable_testing()
+    add_subdirectory(./test)
+endif (GTEST)
 
 set_target_properties ( SLS
     PROPERTIES
