+++ .travis.yml
@@ -19,5 +19,5 @@ before_script:
     - mkdir build
     - cd build
     - cmake ..
-script: make
+script: make && make test
 
