### 2016-12-10 02b33b5
Updated travis cmake version
