+++ .travis.yml
@@ -9,7 +9,10 @@ addons:
     apt:
         sources:
             - ubuntu-toolchain-r-test
+            - george-edison55-precise-backports # cmake 3.2.3 / doxygen 1.8.3
         packages:
+            - cmake
+            - cmake-data
             - gcc-4.8
             - g++-4.8
             - clang
