### 2016-12-10 a2e56c9
Use glm compiled from source
