+++ .travis.yml
@@ -1,3 +1,4 @@
+dist: trusty
 sudo: required
 language: cpp
 compiler:
