### 2016-12-10 d373cd1
Enabled test flag in cmake
