### 2016-12-12 5f75b52
updated readme; removed pointer from App.cpp; Removed anoying error msg
