+++ src/lib/core/Ray.h
@@ -58,7 +58,7 @@ inline glm::vec4 midPointBkp( const Ray &r1, const Ray &r2, float &dist)
     float denom = v1_dot_v1 * v2_dot_v2 - v1_dot_v2 * v1_dot_v2;
     if (glm::abs(denom) < 0.1)
     {
-        LOG::writeLogErr("Par rays\n");
+        // Parallel rays
         dist = -1.0;
         return vec4(0.0);
     }
