+++ README.md
@@ -1,6 +1,6 @@
 # 3DUNDERWORLD-Structured-Light-Scanner
 [![status](http://joss.theoj.org/papers/4329bcbc7bba33961a5e749dcacb995b/status.svg)](http://joss.theoj.org/papers/4329bcbc7bba33961a5e749dcacb995b)
-![build](https://travis-ci.org/theICTlab/3DUNDERWORLD-SLS-GPU_CPU.svg?branch=dev)
+[![Build Status](https://travis-ci.org/theICTlab/3DUNDERWORLD-SLS-GPU_CPU.svg?branch=dev)](https://travis-ci.org/theICTlab/3DUNDERWORLD-SLS-GPU_CPU)
 
 Structured light scanner is a tool to reconstruct point cloud from series of images lit by patterned light. As showing in the following image, patterns are projected to the object in the grid manner. Each cell in the grid has a unique sequence. The corresponding points can be extracted by decoding the sequences from both images.
 ![img](https://raw.githubusercontent.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU/dev/screenshots/flow.png)
@@ -27,19 +27,35 @@ If CUDA is detected in your system, the cmake flag `ENABLE_CUDA` is set to `on` 
 
 You can use `cmake .. -DENABLE_CUDA=off` to disable CUDA if you don't want compile the GPU binary.
 
+### Enable test
+A test using [Google Test Framework](https://github.com/google/googletest.git) is included in the build. To use the test, enable `GTEST` flag at configuration stage.
 ```
-  SLS/SLS_GPU [OPTION...]
+git clone https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU.git
+cd 3DUNDERWORLD-SLS-GPU_CPU
+mkdir build
+cd build
+cmake .. -DGTEST=ON
+make
+make test
+```
+Google test will be downloaded and compiled during the compilation stage. 
 
-  -l, --leftcam arg      Left camera image folder (default:
-                         ../../data/alexander/leftCam/dataset1/)
-  -L, --leftconfig arg   Left camera configuration file (default:
-                         ../../data/alexander/leftCam/calib/output/calib.xml)
-  -r, --rightcam arg     Right camera image folder (default:
-                         ../../data/alexander/rightCam/dataset1/)
-  -R, --rightconfig arg  Right camera configuration file (default:
-                         ../../data/alexander/rightCam/calib/output/calib.xml)
-  -o, --output arg       Output file (default: test.ply)
-  -h, --help             Print this help
+Please note that the test requires reference file from source path, so the CMake output folder needs to be placed in the source path. (Shown as the example above). 
+
+### Run demo binary
+```
+usage: ./SLS [options] ... 
+options:
+  -l, --leftcam        Left camera image folder (string [=../../data/alexander/leftCam/dataset1/])
+  -r, --rightcam       Right camera image folder (string [=../../data/alexander/rightCam/dataset1/])
+  -L, --leftconfig     Left camera configuration file (string [=../../data/alexander/leftCam/calib/output/calib.xml])
+  -R, --rightconfig    Right camera configuration file (string [=../../data/alexander/rightCam/calib/output/calib.xml])
+  -o, --output         Right camera configuration file (string [=output.ply])
+  -f, --format         suffix of image files, e.g. jpg (string [=jpg])
+  -w, --width          Projector width (unsigned long [=1024])
+  -h, --height         Projector height (unsigned long [=768])
+  -?, --help           print this message
+
 ```
 The configuration files are `xml` files generated by OpenCV calibration tool, including intrinsic and extrinsic parameters. 
 
[…]

[…]
