### 2016-12-12 #47 @v3c70r
Sorry for the late reply.
In the latest update, a test has been added and a continuous integration is set up using TravisCI. The test is CPU only for now.
