### 2016-12-20 #47 @daviddoria
Sorry for so much whining about the build process, but I really think we should have a standard of "it just works, or at least there are very explicit instructions". I haven't tried working though this, but below what I'm now getting. Did you intend to provide Findglm.cmake ?

CMake Error at CMakeLists.txt:29 (find_package):
  By not providing "Findglm.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "glm", but
  CMake did not find one.

  Could not find a package configuration file provided by "glm" with any of
  the following names:

    glmConfig.cmake
    glm-config.cmake

---------
Note that I have libglm-dev installed:

doria@doria-home:~/source/3DUNDERWORLD-SLS-GPU_CPU$ sudo apt install libglm-dev
Reading package lists... Done
Building dependency tree
Reading state information... Done
libglm-dev is already the newest version.
The following packages were automatically installed and are no longer required:
  dkms libgsoap4 libntdb1 python-ntdb
Use 'apt-get autoremove' to remove them.
0 upgraded, 0 newly installed, 0 to remove and 64 not upgraded.
