### 2016-12-20 #47 @v3c70r
I installed `glm` from [source code](https://github.com/g-truc/glm/releases/tag/0.9.8.3) and it can be found. It looks like that the `libglm-dev` package doesn't have the config file. I will provide a `findglm.cmake` to cover this situation.
