### 2016-12-20 9eadabf
Modified cmake to search cmake installed by apt
