### 2016-12-20 c6e84a3
Removed data from source
