+++ data/alexander/leftCam/calib/output/cam_distortion.txt
@@ -1,1 +1,0 @@
-0.371032	6.99762	-0.00234272	-0.00266141	-125.243	
