+++ data/alexander/rightCam/calib/output/cam_distortion.txt
@@ -1,1 +1,0 @@
--0.0787473	4.83634	-0.0022975	-0.00294006	-115.112	
