+++ data/alexander/rightCam/calib/output/cam_rotation_matrix.txt
@@ -1,3 +1,0 @@
--0.732851	0.318267	0.601361	
--0.239914	-0.947957	0.209329	
-0.636687	0.00913164	0.771068	
