### 2016-12-20 4fd329e
Removed files
