+++ data/alexander/leftCam/calib/output/cam_matrix.txt
@@ -1,3 +1,0 @@
-12217.4	0	2301.66	
-0	12215.6	1689.8	
-0	0	1	
