### 2016-12-21 2e133d4
Added arch as test data
