+++ test/RunTest.cpp
@@ -67,37 +67,37 @@ bool comparePlyFiles(std::string file1, std::string file2)
     return true;
 }
 
-//TEST( RunCPUTest, Arc)
-//{
-//    const std::string L_IMGS = "../../data/arc/leftCam/dataset1/";
-//    const std::string L_CFG = "../../data/arc/leftCam/calib/output/calib.xml";
-//    const std::string R_IMGS = "../../data/arc/rightCam/dataset1/";
-//    const std::string R_CFG = "../../data/arc/rightCam/calib/output/calib.xml";
-//    const std::string SUFFIX = "tif";
-//    const size_t W=1280, H=800;
-//    const std::string TEST_PLY = "../../test/data/arc.ply";
-//    const std::string TEST_OBJ = "../../test/data/arc.obj";
-//    const std::string O_PLY="arc.ply", O_OBJ="arc.obj";
-//
-//    auto RC = SLS::FileReader("RightCamera");
-//    RC.loadImages(R_IMGS, SUFFIX);
-//    RC.loadConfig(R_CFG);
-//
-//    auto LC = SLS::FileReader("LeftCamera");
-//    LC.loadImages(L_IMGS, SUFFIX);
-//    LC.loadConfig(L_CFG);
-//
-//    SLS::ReconstructorCPU rec(W, H);
-//    rec.addCamera(&LC);
-//    rec.addCamera(&RC);
-//    rec.reconstruct();
-//
-//    SLS::exportPointCloud( O_PLY, "ply", rec);
-//    SLS::exportPointCloud( O_OBJ, "obj", rec);
-//    EXPECT_TRUE(compareObjFiles(TEST_OBJ, O_OBJ));
-//    EXPECT_TRUE(comparePlyFiles(TEST_PLY, TEST_PLY));
-//    EXPECT_TRUE(true);
-//}
+TEST( RunCPUTest, Arch)
+{
+    const std::string L_IMGS = "../../data/arch/leftCam/dataset1/";
+    const std::string L_CFG = "../../data/arch/leftCam/calib/output/calib.xml";
+    const std::string R_IMGS = "../../data/arch/rightCam/dataset1/";
+    const std::string R_CFG = "../../data/arch/rightCam/calib/output/calib.xml";
+    const std::string SUFFIX = "tif";
+    const size_t W=1280, H=800;
+    const std::string TEST_PLY = "../../test/data/arch.ply";
+    const std::string TEST_OBJ = "../../test/data/arch.obj";
+    const std::string O_PLY="arch.ply", O_OBJ="arch.obj";
+
+    auto RC = SLS::FileReader("RightCamera");
+    RC.loadImages(R_IMGS, SUFFIX);
+    RC.loadConfig(R_CFG);
+
+    auto LC = SLS::FileReader("LeftCamera");
+    LC.loadImages(L_IMGS, SUFFIX);
+    LC.loadConfig(L_CFG);
+
+    SLS::ReconstructorCPU rec(W, H);
+    rec.addCamera(&LC);
+    rec.addCamera(&RC);
+    rec.reconstruct();
+
+    SLS::exportPointCloud( O_PLY, "ply", rec);
+    SLS::exportPointCloud( O_OBJ, "obj", rec);
+    EXPECT_TRUE(compareObjFiles(TEST_OBJ, O_OBJ));
+    EXPECT_TRUE(comparePlyFiles(TEST_PLY, TEST_PLY));
+    EXPECT_TRUE(true);
+}
 
 TEST( RunCPUTest, Alexander)
 {
