### 2016-12-21 ce87822
Removed clang from travis
