+++ .travis.yml
@@ -3,7 +3,6 @@ sudo: required
 language: cpp
 compiler:
     - gcc
-    - clang
 install:
     - if [ "$CXX" = "g++" ]; then export CXX="g++-4.8" CC="gcc-4.8"; fi
     - wget https://github.com/g-truc/glm/archive/0.9.8.3.tar.gz
