### 2017-01-04 3506da3
Fixed GPU output for Alexander example
