+++ src/lib/core/Exporter.hpp
@@ -9,6 +9,7 @@ void exportPLYGrid(std::string fileName, const Reconstructor& reconstructor);
 void exportPLY( std::string fileName ,const Reconstructor& reconstructor);
 void exportOBJ( std::string fileName ,const Reconstructor& reconstructor);
 void exportOBJVec4( std::string fileName ,const Reconstructor& reconstructor);
+void dumpRawPointCloud(std::string fileName, const std::vector<float> &pointCloud);
 
 void exportPointCloud(std::string fileName, std::string type, const Reconstructor& reconstructor);
 void exportPointCloud2OBJ(std::string fileName, const std::vector<float> &pointCloud);
