### 2017-01-05 e06c7da
Fixed GPU projector resolution issue
