### 2017-01-10 #47 @daviddoria
I'm not sure I understand the question "Version: Does the release version given match the GitHub release (v4.0.0)?" - there are tons of changes since this release (7 months ago) - shouldn't we just be concerned with whether the branch that is approved here is tagged/referenceable?
