### 2017-01-10 #47 @daviddoria
The test doesn't seem to look for data in the correct place (see ctest output below). Presumably it's from these lines in RunTest.cpp:

const std::string L_IMGS = "../../data/arch/leftCam/dataset1/";

That should probably be something more like:

const std::string L_IMGS = dataPath + "/data/arch/leftCam/dataset1/";

In fact, I'd rather see ctest call the SLS executable the same way a user would from the command line. That is, instead of a special .cpp file (RunTest.cpp), the test would not even need its own add_executable but would just look like:

add_test( NAME CPU_TEST
    COMMAND [path/to/executable/]SLS [options])

It also doesn't make sense to me that the SLS executable is also looking for data in some particular place (see also below). It seems like the path to the data should be a required parameter.

---------------
doria@doria-home:~/build/underworld$ ./bin/SLS
 No image read from ../../data/alexander/rightCam/dataset1/ ...  No image read from ../../data/alexander/leftCam/dataset1/ ... Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Failed to load config ../../data/alexander/rightCam/calib/output/calib.xml
Segmentation fault (core dumped)

---------------

doria@doria-home:~/build/underworld$ ctest -V
UpdateCTestConfiguration  from :/home/doria/build/underworld/DartConfiguration.tcl
UpdateCTestConfiguration  from :/home/doria/build/underworld/DartConfiguration.tcl
Test project /home/doria/build/underworld
Constructing a list of tests
Done constructing a list of tests
Updating test list for fixtures
Added 0 tests to meet fixture requirements
Checking test dependency graph...
Checking test dependency graph end
test 1
    Start 1: CPU_TEST

1: Test command: /home/doria/build/underworld/test/runCPUTest
1: Test timeout computed to be: 9.99988e+06
1: Running main() from gtest_main.cc
1: [==========] Running 2 tests from 1 test case.
1: [----------] Global test environment set-up.
1: [----------] 2 tests from RunCPUTest
1: [ RUN      ] RunCPUTest.Arch
1:  No image read from ../../data/arch/rightCam/dataset1/ ... Failed to load config ../../data/arch/rightCam/calib/output/calib.xml
1: Failed to load config ../../data/arch/rightCam/calib/output/calib.xml
1: Failed to load config ../../data/arch/rightCam/calib/output/calib.xml
1: Failed to load config ../../data/arch/rightCam/calib/output/calib.xml
1/1 Test #1: CPU_TEST .........................***Exception: SegFault  0.15 sec

0% tests passed, 1 tests failed out of 1

Total Test time (real) =   0.16 sec

The following tests FAILED:
          1 - CPU_TEST (SEGFAULT)
Errors while running CTest
