### 2017-01-10 2a23a95
Fixed color export type, color is correct for CPU
