### 2017-01-10 bdb4e82
Removed default data parameters
