### 2017-01-10 #47 @v3c70r
@daviddoria

Thank you for your input,
I have removed the fixed path from default command line parameter.

Regarding the test, since the main contribution of the project are the reconstruction libraries, the main purpose of the test focuses more on the libraries instead of the reconstruction binaries. Also with fixed data path, it would be easier to integrate with the CI.
