### 2017-01-10 #47 @daviddoria
@v3c70r
Sorry, I'm not buying either of those arguments.

> Regarding the test, since the main contribution of the project are the reconstruction libraries, the main purpose of the test focuses more on the libraries instead of the reconstruction binaries.

If you look at App.cpp, it is almost identical to each one of the TEST() in RunTest.cpp . Both very clearly demonstrate the library - there is almost zero overhead in the executable, so I don't know how you think the test would be "testing the binary" instead of "testing the library"?

My point here is avoiding code duplication for maintenance reasons - if you change an API, you will have to update identical code in both App.cpp and RunTest.cpp which just seems unnecessary.

> Also with fixed data path, it would be easier to integrate with the CI.

Those fixed paths imply you are using an in-source build, which is almost always bad practice. I also find it hard to believe that your CI doesn't integrate with CMake well enough to pass a path to the test executable.

My main point is that if I'm a user and run 'make; ctest' and see 100% of tests fail, I don't get a warm fuzzy feeling.
