### 2017-01-11 c2d4cd4
Added external path for test script, added code coverage with gcov
