+++ test/RunTest.cpp
@@ -69,14 +69,14 @@ bool comparePlyFiles(std::string file1, std::string file2)
 
 TEST( RunCPUTest, Arch)
 {
-    const std::string L_IMGS = "../../data/arch/leftCam/dataset1/";
-    const std::string L_CFG = "../../data/arch/leftCam/calib/output/calib.xml";
-    const std::string R_IMGS = "../../data/arch/rightCam/dataset1/";
-    const std::string R_CFG = "../../data/arch/rightCam/calib/output/calib.xml";
+    const std::string L_IMGS = std::string(TEST_DATA_PATH) + "/arch/leftCam/dataset1/";
+    const std::string L_CFG = std::string(TEST_DATA_PATH) + "/arch/leftCam/calib/output/calib.xml";
+    const std::string R_IMGS = std::string(TEST_DATA_PATH) + "/arch/rightCam/dataset1/";
+    const std::string R_CFG = std::string(TEST_DATA_PATH) + "/arch/rightCam/calib/output/calib.xml";
     const std::string SUFFIX = "tif";
     const size_t W=1280, H=800;
-    const std::string TEST_PLY = "../../test/data/arch.ply";
-    const std::string TEST_OBJ = "../../test/data/arch.obj";
+    const std::string TEST_PLY = std::string(TEST_REF_PATH) + "/arch.ply";
+    const std::string TEST_OBJ = std::string(TEST_REF_PATH) + "/arch.obj";
     const std::string O_PLY="arch.ply", O_OBJ="arch.obj";
 
     auto RC = SLS::FileReader("RightCamera");
@@ -96,19 +96,18 @@ TEST( RunCPUTest, Arch)
     SLS::exportPointCloud( O_OBJ, "obj", rec);
     EXPECT_TRUE(compareObjFiles(TEST_OBJ, O_OBJ));
     EXPECT_TRUE(comparePlyFiles(TEST_PLY, TEST_PLY));
-    EXPECT_TRUE(true);
 }
 
 TEST( RunCPUTest, Alexander)
 {
-    const std::string L_IMGS = "../../data/alexander/leftCam/dataset1/";
-    const std::string L_CFG = "../../data/alexander/leftCam/calib/output/calib.xml";
-    const std::string R_IMGS = "../../data/alexander/rightCam/dataset1/";
-    const std::string R_CFG = "../../data/alexander/rightCam/calib/output/calib.xml";
+    const std::string L_IMGS = std::string(TEST_DATA_PATH) + "/alexander/leftCam/dataset1/";
+    const std::string L_CFG = std::string(TEST_DATA_PATH) + "/alexander/leftCam/calib/output/calib.xml";
+    const std::string R_IMGS = std::string(TEST_DATA_PATH) + "/alexander/rightCam/dataset1/";
+    const std::string R_CFG = std::string(TEST_DATA_PATH) + "/alexander/rightCam/calib/output/calib.xml";
     const std::string SUFFIX = "jpg";
     const size_t W=1024, H=768;
-    const std::string TEST_PLY = "../../test/data/alexander.ply";
-    const std::string TEST_OBJ = "../../test/data/alexander.obj";
+    const std::string TEST_PLY = std::string(TEST_REF_PATH) + "/alexander.ply";
+    const std::string TEST_OBJ = std::string(TEST_REF_PATH) + "/alexander.obj";
     const std::string O_PLY="alexander.ply", O_OBJ="alexander.obj";
 
     auto RC = SLS::FileReader("RightCamera");
@@ -128,7 +127,6 @@ TEST( RunCPUTest, Alexander)
[…]
