+++ .travis.yml
@@ -25,6 +25,6 @@ before_script:
     - git submodule update --recursive --remote
     - mkdir build
     - cd build
-    - cmake .. -DGTEST=ON -DENABLE_CUDA=OFF
+    - cmake .. -DGTEST=ON -DENABLE_CUDA=OFF -DTEST_DATA_PATH="../../data/" -DTEST_REF_PATH="../../test/data/"
 script: make && make test
 
