### 2017-01-11 #47 @v3c70r
@daviddoria
As suggested, I have added two mandatory cmake parameters: `TEST_DATA_PATH` and `TEST_REF_PATH` when test is enabled. Now both application and test can be built and run out of source.

> My point here is avoiding code duplication for maintenance reasons - if you change an API, you will have to update identical code in both App.cpp and RunTest.cpp which just seems unnecessary.

I agree with your point here, but I would like to stick with the tests on the library since it will be more flexible. I could extend the test to cover more code and add tests to test only part of the code (e.g. tests on mask ). `App.cpp` contains few code and update it won't be too much trouble (though not quit elegant).
