### 2017-01-11 #47 @arfon
> I agree with your point here, but I would like to stick with the tests on the library since it will be more flexible. I could extend the test to cover more code and add tests to test only part of the code (e.g. tests on mask ). `App.cpp` contains few code and update it won't be too much trouble (though not quit elegant).

This seems reasonable to me. We definitely don't want 'perfect to be the enemy of good' here and the fact that there are _any_ tests puts this software head and shoulders above much of the scientific software online.

@daviddoria - your feedback & input here is much appreciated. In your opinion how close are we to finishing up this review?
