### 2017-01-11 #47 @daviddoria
"We definitely don't want 'perfect to be the enemy of good' here"

Totally agree. I thought this would be an easy fix since it was just de-duplication :)

"and the fact that there are any tests puts this software head and shoulders above much of the scientific software online."

I would philosophically disagree on this one! Just because everyone else is making crap doesn't mean we should lower the bar! "If it's not tested, it's broken" is a pretty solid motto to live by in my opinion.

----------
In the instructions, this:
git clone https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU.git

needs to be changed to
git clone --recursive https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU.git

since the data is now a submodule. You should assume most users will see an empty folder and be very confused (and not know "ah, I should just 'git submodule init; git submodule update' ")

--------------

Guys, can we please try to look at this from a new user's point of view? Why not auto-set TEST_DATA_PATH to ${CMAKE_SOURCE_DIR}/data/whatever ? I don't know the difference between "TEST_DATA_PATH" and "TEST_REF_PATH", and I've been looking at this for a while now. These are not explained in the CMake, and there are no "obvious" folders that one could match to this without instructions (i.e. if there was data/alexander/test and data/alexander/ref)

------------

Unfortunately I'm sensing everyone is starting to get antsy to accept this, while I have barely even started to review the actual code at all! I never even looked in any of the .cpp files, because most of the .h files didn't give me a clue what was going on/what I should be looking for. Many of my comments (particularly about uncommented interfaces) from Oct. 8 are still unaddressed.

I'm happy to step down as a reviewer if you're looking for an "easier" review/reviewer, but I've spent a decade being frustrated with exactly these sorts of issues, so I'm not going to sign off on something that contains them!
