### 2017-01-12 2b1ddfc
Made reconstruct function return a point cloud; Updated readme
