### 2017-01-12 #47 @Kevin-Mattheus-Moerman
As the reviewer guidelines http://joss.theoj.org/about#reviewer_guidelines say
> Reviewers are expected to install the software they are reviewing and to verify the core functionality of the software.

Our guidelines/requirements in relation to tests are at present rather meagre, as it simply says:
> **Tests**
> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
> Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> OK: Documented manual steps that can be followed to check the expected functionality of the software (e.g. a sample input file to assert behaviour)
> Bad (not acceptable): No way for you the reviewer to check whether the software works

Therefore in relation to testing, a simple or sub-optimal tests system may suffice. Having said that, obviously the reviewer should be able to actually set-up the software environment, and it should, through the tests implemented, by easy to assess the core functionality of the software. If the tests are so poorly implemented the latter is not feasible we need to recommend to revise the test suite. If instead the tests are just sub-optimal we may recommend revision but acceptance may in this case (if the core functionality can be evaluated appropriately) not depend on this.

I feel @daviddoria has done a marvellous job at reviewing and I have been very impressed with the speed and level of detail and commitment of his review work. I can assure you we are not “antsy to accept” at this point. From all of the reviewer's comments it seems clear that even as an expert there have been many challenges to overcome to finally get to the point where we can start to evaluate the core functionality of this submission.

@daviddoria thanks again for your efforts on this. This submission is clearly benefiting from the issues you have have raised. You mention that some of the issues raised e.g. around October 8th have not been addressed yet. To avoid confusion, and if it is not too much trouble, would you be so kind to provide an updated list of points that remain to be addressed. Keeping in mind the reviewer guidelines and our current “not so strict” rules on testing methods.

@v3c70r please address all of the points raised by the reviewers. If you disagree on certain aspects clearly detail your arguments against the issues raised and refer to our submission guidelines where possible.

@mikhail-matrosov and @vaheta thanks also to your review contributions. If you are available, it would be much appreciated if you could also pick up the review process as it seems we should now be able to start to get at the core functionality of the software.
