### 2017-01-12 #47 @v3c70r
> Guys, can we please try to look at this from a new user's point of view? Why not auto-set TEST_DATA_PATH to ${CMAKE_SOURCE_DIR}/data/whatever ? I don't know the difference between "TEST_DATA_PATH" and "TEST_REF_PATH", and I've been looking at this for a while now. These are not explained in the CMake, and there are no "obvious" folders that one could match to this without instructions (i.e. if there was data/alexander/test and data/alexander/ref)

I have removed those two unnecessary parameters and use `${CMAKE_SOURCE_DIR}` instead.

Regarding the comment earlier
>From looking at Reconstructor.h, I don't understand the API. What does it mean to add a camera to this object? It's not clear that pointCloud_ (I think) is an output object. I don't know what a Projector is. what is idx_? Why does reconstruct() not return the point cloud?

Now the `reconstruct()` function returns a point cloud and the point cloud can be exported to ply or obj file.

There are still some documentation works to be done but the core function should be working fine now.

I'd like to thank @daviddoria and everybody here who gave me very professional and useful comments, which makes the software more useful and allows me to write better code in the future.
