### 2017-01-13 #47 @daviddoria
Excellent - 'make' now works out of the box, and 'ctest' runs with no setup required and the test passes!

Some more notes (in no particular order):

- everything in test/splitstring should be able to be replaced with a couple of calls to some stdlib functions, no? (or using std::stringstream or something). It's super "old school" as-is.

- I really like how "thin" the files in app/ are!

- In Calibrator.hpp:
  - showImgAvecText_Block - is that Avec=French for "with"? I'd suggest making all function names in English
  - most of the functions do not have descriptions - particularly important is Calibrate()
  - It seems like the UI for selecting the checkerboard corners should be separated from the Calibrate() function, as one could imagine providing their own pixels from another process instead of using this UI

- It seems odd that FileReader derives from Camera? (and following that, that it has functions like undistortPixel, etc.)

- I would provide an explanation in PointCloud.hpp of why you are using a giant float array with all of the values from all of the points instead of the expected vector<Point> or similar

- The documentation for Projector doesn't explain what it does, and presumably it's a core component? (though it doesn't seem to do really anything related to "projection"?)

- GrayCode.hpp has zero documentation

- I would add some "why do we need this" motivation to DynamicBitset. I wonder if it's worth a Boost dependency for boost::dynamic_bitset? It seems like a lot of highly error-prone code in there.

- In ReconstructorCPU.h:
  - What are "buckets"? They seem to be a core concept, but are not explained at all
  - If you could break the two big functions down into 3 or 4 each, I imagine the code would be much easier to follow. Good places to start are with the loops - whatever is happening each time through the loop is typically a very easy thing to extract and name. Then by looking at the class interface, the reader can know all of the things that happen, and by looking at the implementation they just learn the order in which those things happen. It also allows a much more "hey, I could change/customize this to try out my new idea!" attitude, instead of "gah, this function is huge, I'll probably break something if I touch it"

- In the README, I would replace the copy/pasted code (#1: if it's not compiled, it's broken even more than if it's not tested, and #2, you should just be able to point the reader to one of the real files that exists already) with a high level system description. That is, something like (but hopefully more involved) "The implementation is very straight forward. We construct a collection of Camera objects, and use the Reconstructor to them to do XYZ with Buckets (which definitely seems like it wants to be a real object) which finally produces a PointCloud of the reconstruction".
