### 2017-01-13 #47 @v3c70r
Thanks @daviddoria , I'm starting to work on the issues now.

> I would add some "why do we need this" motivation to DynamicBitset. I wonder if it's worth a Boost dependency for boost::dynamic_bitset? It seems like a lot of highly error-prone code in there.

I take a quick look at the boost::dynamic_bitset. It looks like exactly the library I want.

>The documentation for Projector doesn't explain what it does, and presumably it's a core component? (though it doesn't seem to do really anything related to "projection"?)

The projector is literally a projector that projects the pattern on the object for reconstruction. I'll add more details in the documentation.

> It seems odd that FileReader derives from Camera? (and following that, that it has functions like undistortPixel, etc.)

Here the camera is defined as a reconstruction input device. It encapsulates image acquisition, loading configuration, undistortion, and masking out invalid pixels. `FileReader` acquires images and configurations from file, which satisfies these interfaces. I'm thinking this class probably need a better name like `DataInput`?

> I would provide an explanation in PointCloud.hpp of why you are using a giant float array with all of the values from all of the points instead of the expected vector or similar

I think when deal with the GPU implementation, it would be easier to pass in primitive data types.
