### 2017-01-13 03d49ee
Removed splitstring from dependency
