### 2017-01-13 9fba710
removed splitstring from dependency
