+++ test/RunTest.cpp
@@ -1,7 +1,7 @@
 #include <fstream>
 #include <core/FileReader.h>
 #include <core/ReconstructorCPU.h>
-#include "splitstring.h"
+#include <sstream>
 #include <gtest/gtest.h>
 
 const float MAX_DIFF = 0.5;
@@ -11,31 +11,30 @@ inline bool comparePlyLine(std::string line1, std::string line2)
     if (!std::isdigit(line1[0]) && !std::isdigit(line1[1]))    // if is header
         return line1 == line2;
 
-    splitstring split1(line1);
-    splitstring split2(line2);
-    vector<std::string> splited1 = split1.split(' ');
-    vector<std::string> splited2 = split2.split(' ');
-    if (splited1.size() > 0 && splited2.size() > 0)
-    {
-        for (size_t i=0 ; i<splited1.size(); i++)
-            if ( std::fabs( std::stof(splited1[i]) - std::stof(splited2[i]) ) > MAX_DIFF)
-                return false;
-    }
+    std::istringstream ssline1(line1);
+    std::istringstream ssline2(line2);
+    std::string num1, num2;
+
+    while (std::getline(ssline1, num1, ' ') && std::getline(ssline2, num2, ' '))
+        if ( std::fabs( std::stof(num1) - std::stof(num2) ) > MAX_DIFF)
+            return false;
+
     return true;
 }
 
 inline bool compareObjLine(std::string line1, std::string line2)
 {
-    splitstring split1(line1);
-    splitstring split2(line2);
-    vector<std::string> splited1 = split1.split(' ');
-    vector<std::string> splited2 = split2.split(' ');
-    if (splited1.size() > 0 && splited2.size() > 0)
+    std::istringstream ssline1(line1);
+    std::istringstream ssline2(line2);
+    std::string num1, num2;
+
+    if (std::getline(ssline1, num1, ' ') && std::getline(ssline2, num2, ' ')) // Skip first split
     {
-        for (size_t i=1 ; i<splited1.size(); i++)
-            if ( std::fabs( std::stof(splited1[i]) - std::stof(splited2[i]) ) > MAX_DIFF)
+        while (std::getline(ssline1, num1, ' ') && std::getline(ssline2, num2, ' '))
+            if ( std::fabs( std::stof(num1) - std::stof(num2) ) > MAX_DIFF)
                 return false;
     }
+
     return true;
 }
 
