### 2017-01-13 #47 @daviddoria
> Here the camera is defined as a reconstruction input device. It encapsulates image acquisition, loading configuration, undistortion, and masking out invalid pixels. FileReader acquires images and configurations from file, which satisfies these interfaces. I'm thinking this class probably need a better name like DataInput?

I guess the point was that use of a hierarchy seems odd. I'd expect something more like a FileReader/DataInput object to take a file/folder and return a Camera. Having FileReader *derive* from camera doesn't seem like the right model.

I guess I'm also confused that the class that loads data has a function called 'undistortPixel', unless you're considering part of loading data also "preparing it" in the sense that you're removing distortion from the images you load. This part can be fixed with a few sentences in a class description :)

The getRay functions, computeShadows.., etc. also seem like they should just be directly part of the Camera class.

> I think when deal with the GPU implementation, it would be easier to pass in primitive data types.

I assumed that was the case, but it should still be explained in the class to avoid confusion.

---
One new thing - the variable "color_" in Camera seems like it should actually be called "litImage_" or similar? I hope you can see I'm not trying to be obnoxious and "perfect" names (taking them from "pretty good" to "pretty good++"), but rather trying to take them from "non-descriptive/very confusing" to "readable/comprehensible".
