### 2017-01-15 0d26aa5
Let calibrator take parameters, added more parameters in loadImage() function
