### 2017-01-16 5a934c1
Trying to add doc to CMake
