+++ doc/CMakeLists.txt
@@ -1,0 +1,3 @@
+find_package(Doxygen REQUIRED)
+message(STATUS ${DOXYGEN_EXECUTABLE})
+
