### 2017-01-18 9850ba2
Modified documentation
