+++ doc/doxygen.in
@@ -1,12 +1,19 @@
 PROJECT_NAME           = "@CMAKE_PROJECT_NAME@"
 PROJECT_NUMBER         = @VERSION_MAJOR@.@VERSION_MINOR@.@VERSION_PATCH@
-STRIP_FROM_PATH        = @PROJECT_SOURCE_DIR@ \
-                         @PROJECT_BINARY_DIR@
+STRIP_FROM_PATH        = @PROJECT_SOURCE_DIR@/src/lib
 INPUT                  = @doxy_main_page@ \
-                         @PROJECT_SOURCE_DIR@ \
-                         @PROJECT_BINARY_DIR@
+                         @PROJECT_SOURCE_DIR@/src/lib
+
+EXTRACT_ALL            = YES
+EXTENSION_MAPPING      = cu=c++ \
+                         cuh=c++
 FILE_PATTERNS          = *.h \
-                         *.cc
+                         *.cc \
+                         *.hpp \
+                         *.cpp \
+                         *.cu \
+                         *.cuh
+
 RECURSIVE              = YES
 USE_MDFILE_AS_MAINPAGE = @doxy_main_page@
 
