### 2017-01-19 54aa191
added mathjax support in doc; documented PointCloud and Projector.h
