+++ src/lib/GrayCode/GrayCode.hpp
@@ -4,25 +4,23 @@
 #include <cmath>
 
 namespace SLS{
-class GrayCode: public Projector
+class GrayCode
 {
 public:
+    //! Construct gray code with size of projector
     explicit GrayCode(size_t projW, size_t projH);
-    ~GrayCode() override
+    ~GrayCode() 
     {
         for (auto &img: grayCodes_)
             img.release();
     }
     void generateGrayCode();
 private:
+    //! Projector width and height
+    size_t width_, height_;
     std::vector<cv::Mat> grayCodes_;
     size_t currImg;
     size_t numColBits, numRowBits;
-    void setColRowBitsNum()
-    {
-        numColBits = (size_t)std::ceil(std::log2((float) width_));
-        numRowBits = (size_t)std::ceil(std::log2((float) height_));
-    }
 };
 } // SLS
 
