### 2017-01-20 d11fef2
More comment, less dead code
