### 2017-01-23 #47 @charalambos
Hi all,

I guess this conclude the review? How shall we proceed?

Best,

C.
