### 2017-01-23 #47 @v3c70r
@charalambos I'm still addressing the issues. Here is the [progress](https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU/projects/2)
