### 2017-01-26 #47 @Kevin-Mattheus-Moerman
@vaheta That is fine. It may take the authors some time to work on some of the issues discussed above so when you join in you will hopefully be able to work with the revised/updated version.
