### 2017-02-06 ac30183
Fixing build, refactoring reconstructor CPU
