+++ src/lib/core/ReconstructorCPU.h
@@ -1,68 +1,71 @@
-/* Copyright (C) 
- * 2016 - Tsing Gu
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
- * 
- */
 #pragma once
-#include "Reconstructor.h"
-#include "Log.hpp"
-#include <memory>
 #include <core/PointCloud.hpp>
-namespace SLS
-{
+#include <memory>
+#include "Log.hpp"
+#include "Reconstructor.h"
+namespace SLS {
 
 // CPU reconstructor
-class ReconstructorCPU: public Reconstructor
-{
+/*! Reconstruction buckets of cameras
+ *
+ * In the reconstruction, camera pixels are assigned to different projector
+ * pixels.
+ * Generally, one projector pixels contains more than one camera pixel. For each
+ * camera, we assign
+ * pixels to projector pixels, and call those projector pixels buckets.
+ * ```
+ * ProjPixels(Buckets)   Camera pixels
+ * +------------------+  +---+---+---+
+ * |       0          +->+   |   |   |
+ * +------------------+  +-------+---+
+ * |       1          +->+   |
+ * +------------------+  +----
+ *
+ * +------------------+  +---+---+
+ * |       n          +->+   |   |
+ * +------------------+  +---+---+
+ *
+ * ```
+ * The camera pixels in the same bucket of two different cameras are considered
+ * correspondent pixels,
+ * depth then can be extracted from those pixels.
+ */
+class ReconstructorCPU : public Reconstructor {
 private:
-    /*! Reconstruction buckets of cameras
-     *
-     * In the reconstruction, camera pixels are assigned to different projector pixels. 
-     * Generally, one projector pixels contains more than one camera pixel. For each camera, we assign 
-     * pixels to projector pixels, and call those projector pixels buckets.
-     * ```
-     * ProjPixels(Buckets)   Camera pixels
-     * +------------------+  +---+---+---+
-     * |       0          +->+   |   |   |
-     * +------------------+  +-------+---+
-     * |       1          +->+   |
-     * +------------------+  +----
-     * 
-     * +------------------+  +---+---+
-     * |       n          +->+   |   |
-     * +------------------+  +---+---+
-     *
-     * ```
-     * The camera pixels in the same bucket of two different cameras are considered correspondent pixels,
-     * depth then can be extracted from those pixels.
-     */
-    std::vector<std::vector<std::vector<size_t>>> buckets_;//[camIdx][ProjPixelIdx][camPixelIdx]
+    std::vector<std::vector<std::vector<size_t>>>
+        buckets_;  //!< Buckets [camIdx][ProjPixelIdx][camPixelIdx]
     void initBuckets();
     void generateBuckets();
-public:
 
+    //! Find intersection of two projector pixel
+    /*! Each projector pixel contains multiple camera pixels.
+     * This function takes two buckets from two cameras. Shoot a ray from each
+     * pixel to find a pair of rays with minimum distance. The midpoints of the
+     * closest rays are considered intersection points.
+     *
+     * The retrun value is put in a std::array<glm::vec3, 2>, in which the first
[…]
