+++ src/lib/core/ReconstructorCPU.cpp
@@ -1,36 +1,28 @@
-#include <iomanip>
 #include "ReconstructorCPU.h"
-#include "FileReader.h"
+#include <iomanip>
 #include <limits>
+#include "FileReader.h"
 
-namespace SLS{
+namespace SLS {
 
-ReconstructorCPU::~ReconstructorCPU()
-{
-    delete projector_;
-}
+ReconstructorCPU::~ReconstructorCPU() { delete projector_; }
 void ReconstructorCPU::initBuckets()
 {
-    size_t x,y;
-    projector_->getSize(x,y);
+    size_t x, y;
+    projector_->getSize(x, y);
     buckets_.resize(cameras_.size());
-    for (auto& b: buckets_)
-        b.resize(x*y);
+    for (auto &b : buckets_) b.resize(x * y);
     generateBuckets();
 }
 
-void ReconstructorCPU::addCamera(Camera *cam)
-{
-    cameras_.push_back(cam);
-}
-
+void ReconstructorCPU::addCamera(Camera *cam) { cameras_.push_back(cam); }
 void ReconstructorCPU::generateBuckets()
 {
-    //Generating reconstruction bucket for each camera
-    //for each camera
+    // Generating reconstruction bucket for each camera
     for (size_t camIdx = 0; camIdx < cameras_.size(); camIdx++) {
         const auto &cam = cameras_[camIdx];
-        LOG::writeLog("Generating reconstruction bucket for \"%s\" ... \n", cam->getName().c_str());
+        LOG::writeLog("Generating reconstruction bucket for \"%s\" ... \n",
+                      cam->getName().c_str());
 
         cam->computeShadowsAndThresholds();
 
@@ -40,19 +32,21 @@ void ReconstructorCPU::generateBuckets()
 
         // For each camera pixel
         for (size_t i = 0; i < xTimesY; i++) {
-            if (!cam->queryMask(i)) // No need to process if in shadow
+            if (!cam->queryMask(i))  // No need to process if in shadow
                 continue;
 
+            // First two frames are lit and dark frame
+            // not considered
             cam->getNextFrame();
-            cam->getNextFrame(); //skip first two frames
+            cam->getNextFrame();  // skip first two frames
 
             Dynamic_Bitset bits(projector_->getRequiredNumFrames());
 
             bool discard = false;
 
             // for each frame
-            for (int bitIdx = projector_->getRequiredNumFrames() - 1; bitIdx >= 0; bitIdx--) {
-
+            for (int bitIdx = projector_->getRequiredNumFrames() - 1;
+                 bitIdx >= 0; bitIdx--) {
                 auto frame = cam->getNextFrame();
                 auto invFrame = cam->getNextFrame();
                 unsigned char pixel = frame.at<uchar>(i % y, i / y);
@@ -60,77 +54,140 @@ void ReconstructorCPU::generateBuckets()
 
                 // Not considering shadow mask. But the following test should be
                 // more strict than shadow mask.
-                if (invPixel > pixel && invPixel - pixel >= ((FileReader*)cam)->getWhiteThreshold(i)) {
-                    // No need to do anything since the array is initialized as all zeros
-                    bits.clearBit((size_t) bitIdx);
+                if (invPixel > pixel &&
+                    invPixel - pixel >=
+                        ((FileReader *)cam)->getWhiteThreshold(i)) {
+                    // No need to do anything since the array is initialized as
+                    // all zeros
+                    bits.clearBit((size_t)bitIdx);
                     continue;
                 }
-                else if (pixel > invPixel && pixel - invPixel > 
-                        ((FileReader*)cam)->getWhiteThreshold(i)) {
-                    bits.setBit((size_t )bitIdx);
-
[…]
