### 2017-02-07 19a18f7
Added side bar to doxygen
