+++ src/lib/core/DynamicBitset.h
@@ -33,7 +33,7 @@ namespace SLS
 {
 class Dynamic_Bitset{
 private:
-    std::vector<unsigned char> bits;
+    std::vector<unsigned char> bits; //!< Bytes used to store bits
 protected:
     const size_t BITS_PER_BYTE;
 
@@ -42,41 +42,35 @@ protected:
      * http://stackoverflow.com/questions/47981/how-do-you-set-clear-and-toggle-a-single-bit-in-c-c
      */
 
-    /**
-     * @brief Set bit of a char to 1
+    /*! Set bit of a char to 1
      *
-     * @param ch Char to operate
-     * @param pos Position within the char
+     * \param ch Char to operate
+     * \param pos Position within the char
      */
     void setUChar(unsigned char& ch, const size_t &pos) { ch |= 1<<pos; }
-    /**
-     * @brief Set bit of a char to 0
+    /*! Set bit of a char to 0
      *
-     * @param ch Char to operate
-     * @param pos Position within the char
+     * \param ch Char to operate
+     * \param pos Position within the char
      */
     void clearUChar(unsigned char& ch, const size_t &pos) { ch &= ~(1<<pos);}
 
-    /**
-     * @brief Get bit of within a char
+    /*! Get bit of within a char
      *
-     * @param ch Char to query
-     * @param pos position of bit in char
+     * \param ch Char to query
+     * \param pos position of bit in char
      *
-     * @return Ture if 1; otherwise, 0.
+     * \return Ture if 1; otherwise, 0.
      */
     bool getUChar(const unsigned char& ch, const size_t &pos)const {return (ch>>pos)&1;}
 public:
-
-    /**
-     * @brief Init an empty bitset
+    /*! Init an empty bitset
      */
     Dynamic_Bitset(): BITS_PER_BYTE{8}{}
 
-    /**
-     * @brief Init an empty bitset with given length
+    /*! Init an empty bitset with given length
      *
-     * @param sz number of bits 
+     * \param sz number of bits 
      */
     explicit Dynamic_Bitset(const size_t &sz):BITS_PER_BYTE{8}{resize(sz);}
     Dynamic_Bitset(const size_t &sz, unsigned char* b):BITS_PER_BYTE{8}
@@ -88,27 +82,24 @@ public:
         memcpy( &bits[0], b, sz);
     }
 
-    /**
-     * @brief Get number of bits
+    /*!  Get number of bits
      *
-     * @return number of bits
+     * \return number of bits
      */
     virtual size_t size() const {return bits.size()*BITS_PER_BYTE;}
 
-    /**
-     * @brief Resize bit array length, will set bit array to 0
+    /*! Resize bit array length, will set bit array to 0
      *
-     * @param sz number of bits
+     * \param sz number of bits
      */
     virtual void resize(const size_t &sz) {
         bits.resize((sz+BITS_PER_BYTE-1)/ BITS_PER_BYTE, 0);
         memset(&(bits[0]), 0, sizeof(unsigned char)*bits.size());
     }
 
-    /**
-     * @brief Set a bit to 1
+    /*! Set a bit to 1
      *
-     * @param pos 0-based position of bit
+     * \param pos 0-based position of bit
      */
     void setBit(const size_t &pos){    //Set the bit to one
         if (pos > size())
@@ -116,10 +107,9 @@ public:
         setUChar(bits[pos/BITS_PER_BYTE], pos%BITS_PER_BYTE);
     }
 
-    /**
-     * @brief Set a bit to 0
+    /*! Set a bit to 0
      *
-     * @param pos 0-based position of bit
+     * \param pos 0-based position of bit
      */
[…]
