### 2017-02-10 #47 @Kevin-Mattheus-Moerman
@charalambos I noticed that on your list here https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU/projects/2 several points are labelled as done. Let us know when the review process can be resumed.
