### 2017-02-11 #47 @daviddoria
Hi guys,

- Great, the build+test now works out of the box!

- The ply files produced still crash Paraview 5.2 (though it can open the obj files properly).

- I get "DOI not found" for this: http://dx.doi.org/10.21105/joss.00047 which appears in the pdf

- I would make the link to this paper https://arxiv.org/pdf/1406.6595.pdf VERY VERY PROMINENT. Without it, it's nearly impossible to figure out what is going on. I would say right at the very top "this is the source code for the structured light implementation described in detail here: ..."

- In general, I'd still like to see better separation of the components for reusability purposes. For example, in GrayCode.*, the generation of the gray code images is done in the same function that displays them. If I wanted to take this code and do something else with the images (write them to a file, etc.), then I'd have to tear it apart. This also is critically important to write testable code.

Going down the list of things to review, I'm confused about the difference between paper.md and README.md. They both seem to be trying to say the same thing, but paper.md is essentially a weird subset of paper.md. Can we have just one (the larger one). The pdf that is generated is mighty sparse to call a "software paper". I would have like to see something like a class diagram or at least a description of the major system components and a description of the workflow (something like "use X to load data, then construct a Y that produces a Z").

"A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?"

No - I don't see anything about why this is necessary (are there no other open source structured light scanners?). "Who is the audience?" is also not addressed - is this for people that want a implementation to build extensions upon? To actually use to create models in production? To use for "research purposes" (read: it probably doesn't work that well), etc. Is it the GPU implementation that makes this new/interesting?

Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

Yes, but especially since there is only one it should be explained. That is, "run 'make test' and you will see the system transform the "Arch" and "Alexander" datasets from a collection of images into a point cloud. You will find the output in [build]/test/alexander.[ply|obj] and [build]/test/arch.[ply|obj]."

"Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I don't see any of these things listed. Presumably since it's on github I would imagine pull requests are welcome. Should we email the authors (I don't see any email addresses) or create an issue on github if we have a question?, etc.

"A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?"

No - same as above. I'm still not sure why this is asked again here (which is probably what led to the "duplicate" documentation/file).

"References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?"

There are no references. I would really like to see some pointers to the "canonical" structured light paper(s) and any existing implementations.

More comments from another pass through the code:

- All 3 files in app/ do not have sufficient explanation. For example I ran 'generateGraycode' and was greeted with a white window... after waiting a while I pressed 'q' to try to quit and then the window turned black. I pressed it again and it split black/white. I then realized it was stepping through graycode targets. This should be explained so I don't have accidentally figure out what to do.
For CalibrateCamera.cpp - it should be explained what kind of files are expected in the 'images' directory (pairs? numbers required? etc.) and what kind of calibration technique is used/how the parameters are stored in 'output'. After continuing to look, it could at least say "see Calibrator.hpp for input/output requirements/details."

- I saw using boost::dynamic_bitset is in "not now" on your todo list, which is fine, but then DynamicBitset.h needs at least a short explanation of what it is/does (perhaps with a pointer to "if you're familiar with boost::dynamic_bitset, this does approximately the same thing"

- The FileReader class still doesn't have an explanation. This is particularly confusing because it seems to do much more than read files (especially since it derives from Camera). Also, things like 'getParams()' having documentation of "get parameters" is not helpful at all, especially when the 'params_' member has no explanation.

- Log has no explanation. I see a function like 'progress; // display progress bar' and wonder if going to do some UI sort of thing, CLI ####... type of progress indication, etc.

Thanks,

David
