### 2017-02-12 #47 @arfon
> - I get "DOI not found" for this: http://dx.doi.org/10.21105/joss.00047 which appears in the pdf

@daviddoria - don't worry about that for now. This is the DOI that the publication _will eventually_ have.
