### 2017-02-12 #47 @Kevin-Mattheus-Moerman
@daviddoria Thanks for your hard work!

About readme.md and paper.md. The paper.md is for JOSS and forms the paper document. The paper itself can be quite minimal and might have some overlap with the readme but should reference works properly though.
See here for guidelines on what it should contain: http://joss.theoj.org/about#author_guidelines
