### 2017-02-12 66a5ed2
Separated graycode generation and displaying.
