### 2017-03-13 #47 @v3c70r
Thank you for reminding me of the update. Sorry I haven't worked to much on the fixes these days because of tight schedule and Hackathon weekend. I will pick it up this week and hopefully it would be done by the weekend. Thank you for your patience.
