+++ src/app/App.cpp
@@ -9,6 +9,7 @@
 int main(int argc, char** argv)
 {
     
+    // Parsing parameters
     cmdline::parser p;
     p.add<std::string>("leftcam", 'l',"Left camera image folder", true);
     p.add<std::string>("rightcam", 'r',"Right camera image folder", true);
@@ -20,33 +21,49 @@ int main(int argc, char** argv)
     p.add<size_t>("height", 'h',"Projector height", true);
     p.parse_check(argc, argv);
 
+    // Start logging, this function will clear log file and start a new logging.
     LOG::restartLog();
+
+    // Setup left camera image folder
     std::string leftCameraFolder = p.get<std::string>("leftcam");
-    std::string rightCameraFolder = p.get<std::string>("rightcam");
+    // Setup left camera calibration parameters
     std::string leftConfigFile = p.get<std::string>("leftconfig");
+
+    // Setup right camera
+    std::string rightCameraFolder = p.get<std::string>("rightcam");
     std::string rightConfigFile = p.get<std::string>("rightconfig");
+
     std::string output = p.get<std::string>("output");
     std::string suffix = p.get<std::string>("format");
 
-
+    // Initialize two file readers to load images from file
     SLS::FileReader rightCam("rightCam");
     SLS::FileReader leftCam("rightCam");
 
+    // Load images
+    // void loadImages( const std::string &folder, std::string prefix, size_t numDigits, size_t startIdx, std::string suffix )
     rightCam.loadImages(rightCameraFolder, "", 4, 0,suffix);
     leftCam.loadImages(leftCameraFolder, "", 4, 0, suffix);
 
+    // Load configurations, mainly calibration parameters
     rightCam.loadConfig(rightConfigFile);
     leftCam.loadConfig(leftConfigFile);
 
+    // Initialize a reconstructor with the resolution of the projection to project patterns
     SLS::ReconstructorCPU reconstruct(p.get<size_t>("width"), p.get<size_t>("height"));
+
+    // Add cameras to reconstructor
     reconstruct.addCamera(&rightCam);
     reconstruct.addCamera(&leftCam);
 
+    // Run reconstructio and get the point cloud
     auto pointCloud = reconstruct.reconstruct();
 
+    // Get extension of output file
     auto extension = output.substr(output.find_last_of(".")+1);
+
+    // Export point cloud to file
     pointCloud.exportPointCloud( p.get<std::string>("output"), extension);
-    //SLS::exportPointCloud(output, extension, reconstruct);
 
     LOG::writeLog("DONE!\n");
     
