+++ src/lib/core/Camera.cpp
@@ -1,4 +1,0 @@
-#include "Camera.h"
-namespace SLS{
-}
-
