+++ src/lib/core/ImageProcessor.cpp
@@ -1,0 +1,4 @@
+#include "ImageProcessor.h"
+namespace SLS{
+}
+

[…]
