### 2017-03-29 52fafed
Started refactoring reconstruction process
