+++ src/lib/core/ImageProcessor.cpp
@@ -1,4 +1,4 @@
 #include "ImageProcessor.h"
 namespace SLS{
+ImageProcessor::~ImageProcessor() {}
 }
-
