+++ src/lib/core/ImageFileProcessor.h
@@ -91,5 +91,6 @@ public:
     void computeShadowsAndThresholds() override;
     void setResolution (const size_t &x, const size_t &y) override {resX_ = x; resY_ = y; rayTable.resize(resX_*resY_);}
     unsigned char getWhiteThreshold(size_t i) const { return thresholds_[i];}
+    Buckets generateBuckets(size_t projWidth, size_t projHeight, size_t requiredNumFrames) override;
 };
 }
