### 2017-03-31 23a06f1
Refactored reconstructor, Pending doc changes
