### 2017-04-03 5982af6
Updated readme, finishing doc
