### 2017-04-03 #47 @v3c70r
Sorry for making you wait.
I'm refactoring the code to change the `Camera` into `ImageProcessor` and moved it out of `Reconstructor`. It has been done on the CPU and I still need to do the same change on GPU code. As well as the documentation. The updated pipeline is stated in the [graph of README](https://github.com/theICTlab/3DUNDERWORLD-SLS-GPU_CPU/blob/dev/doc/how-it-works.svg).
