### 2017-11-14 #47 @Kevin-Mattheus-Moerman
@arfon the authors (@charalambos and @v3c70r ) have indictated (by e-mail to me) that they wish to withdraw this submission. Can you process the withdrawal please?
