### 2017-11-14 #47 @Kevin-Mattheus-Moerman
@daviddoria thank you for your time and effort and continued involvement in this review process! Thank you also @vaheta @mikhail-matrosov for your review contributions.
